% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_CRIM}
\alias{Barretts_CRIM}
\title{CRIM status Barrett's}
\usage{
Barretts_CRIM(x, HospNum, EVENT)
}
\arguments{
\item{x}{The dataframe}

\item{HospNum}{The Hospital Number column}

\item{EVENT}{The column  called EVENT that determines what procedure the
patient had at that endoscopy}
}
\description{
This collects the patients in whom it is assumed that complete clearance of 
intestinal metasplasia has occurred (CRIM) after ablation therapy. 
This is done by collecting those endoscopies where the last EVENT was 
equal to 'nothing' when the patient had undergone radiofrequency ablation at 
some point
}
\examples{
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Natureofspecimen','specimen')
v<-HistolChopperBxSize(v,'Natureofspecimen')
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
                             'ProcedurePerformed','OGDReportWhole','Findings')
colnames(b3)[colnames(b3) == 'pHospitalNum'] <- 'HospitalNumber'
Barretts_CRIM(b3,'HospitalNumber',"EVENT")
}
\keyword{CRIM}
