% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_EventType}
\alias{Barretts_EventType}
\title{Therapeutic subtypes}
\usage{
Barretts_EventType(dataframe, HistolReportColumn, ProcPerformedColumn,
  EndoReportColumn, EndoFindingsColumn)
}
\arguments{
\item{dataframe}{the dataframe}

\item{HistolReportColumn}{The histology main text}

\item{ProcPerformedColumn}{The Procedure Performed column}

\item{EndoReportColumn}{The endoscopic diagnosis column}

\item{EndoFindingsColumn}{The endoscopic findings column if different 
to the Diagnosis column}
}
\description{
This function extracts the Event- usually a therapeutic event, from the text
eg endoscopic mucosal resection, radiofrequency ablation etc.
It does not currently include stricture
dilatation.Specfically it extracts the event
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.

v<-HistolAccessionNumber(Mypath,'Histology',
'SP-\\\\d{2}-\\\\d{7}')
v<-HistolDx(v,'Diagnosis')
v<-HistolExtrapolDx(v,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The function then looks within the Histology and the
# Procedure performed, free text endoscopic Findings and the original
# whole endoscopy report columns from the merged data set (v) for
# EMR/APC/RFA/HALO so that the event for the procedure, is recorded.
b<-Barretts_EventType(v,'Histology',
'ProcedurePerformed','Indications','Findings')
rm(v)

}
\keyword{Event}
\keyword{extraction}
