% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_FUType}
\alias{Barretts_FUType}
\title{Follow up group determination}
\usage{
Barretts_FUType(dataframe, CStage, MStage, IMorNoIM)
}
\arguments{
\item{dataframe}{the dataframe(which has to have been processed by the
Barretts_PathStage function first to get IMorNoIM and the Barretts_PragueScore
to get the C and M stage if available),}
}
\description{
This determines the follow up rule a patient should fit in to (according to
the British Society for Gastroenterology guidance on Barrett's oesophagus)
Specfically it combines the presence of intestinal metaplasia with
Prague score so the follow-up group can be determined. It relies on the
presence of a Prague score. It should be run after
Barretts_PathStage which looks for the worst stage of a
specimen and which will determine the presence or absence of intestinal
metaplasia if the sample is non-dysplastic. Because reports often do not record
a Prague score a more pragmatic approach as been to assess the M stage and if 
this is not present then to use the C stage extrapolated using the 
Barretts_Prague function
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolDx(Mypath,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
#The function relies on the other Barrett's functions being run as well:
b<-Barretts_PathStage(v,'Histology')
b2<-Barretts_PragueScore(b,'Findings')
#The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
cc<-Barretts_FUType(b2,"CStage","MStage","IMorNoIM")
rm(v)
}
\keyword{Follow-Up}
