% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_LesionRecognitionEMR}
\alias{Barretts_LesionRecognitionEMR}
\title{Barretts_LesionRecognitionEMR}
\usage{
Barretts_LesionRecognitionEMR(EndoSubsetEMR, y, z)
}
\arguments{
\item{EndoSubsetEMR}{The dataframe}

\item{y}{Endoscopy report field of interest}

\item{z}{Another endoscopy report field of interest}
}
\description{
This looks at the Paris grades of each EMR and then creates a heatmap of pathological grade vs 
endoscopic Paris grade
}
\examples{
Barretts_LesionRecognitionEMR(EndoSubsetEMR,'FINDINGS','DIAGNOSIS')
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
