% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_OPCS4}
\alias{Barretts_OPCS4}
\title{Barrett's OPCS-4 Coding}
\usage{
Barretts_OPCS4(dataframe, EVENT)
}
\arguments{
\item{dataframe}{the dataframe}

\item{EVENT}{the EVENT column}
}
\description{
This function extracts the OPCS-4 codes for all Barrett's procedures
It should take the OPCS-4 from the EVENT and perhaps also using extent
depending on how the coding is done. The EVENT column will need to 
extract multiple findings
The hope is that the OPCS-4 column will then map from the EVENT column
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.

# EXAMPLE TO BE WRITTEN v<-HistolAccessionNumber(Mypath,'Histology')
}
\keyword{OPCS-4}
\keyword{codes}
\keyword{extraction}
