% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Barretts.R
\name{BarrettsAll}
\alias{BarrettsAll}
\title{Run all the basic Barrett's functions}
\usage{
BarrettsAll(Endodataframe, EndoReportColumn, EndoReportColumn2,
  Pathdataframe, PathColumn)
}
\arguments{
\item{Endodataframe}{endoscopy dataframe of interest}

\item{EndoReportColumn}{Endoscopy report field of interest as a string vector}

\item{EndoReportColumn2}{Second endoscopy report field of interest as a string vector}

\item{Pathdataframe}{pathology dataframe of interest}

\item{PathColumn}{Pathology report field of interest as a string vector}
}
\value{
Newdf
}
\description{
Function to encapsulate all the Barrett's functions together. This includes the Prague
score and the worst pathological grade and then feeds both of these things into
the follow up function. The output is a dataframe with all the original data as
well as the new columns that have been created.
}
\examples{
Barretts_df <- BarrettsAll(Myendo, "Findings", "OGDReportWhole", Mypath, "Histology")
}
\seealso{
Other Disease Specific Analysis - Barretts Data: \code{\link{BarrettsBxQual}},
  \code{\link{BarrettsParisEMR}},
  \code{\link{Barretts_FUType}},
  \code{\link{Barretts_PathStage}},
  \code{\link{Barretts_PragueScore}}
}
\concept{Disease Specific Analysis - Barretts Data}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
