% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsBasicNumbers}
\alias{BarrettsBasicNumbers}
\title{BarrettsBasicNumbers}
\usage{
BarrettsBasicNumbers(x, Endo_ResultPerformed)
}
\arguments{
\item{x}{the dataframe}

\item{Endo_ResultPerformed}{the date the endoscopy was performed}
}
\description{
This looks at the basic numbers of all the therapeutic endoscopies over time.Need to have run
BarettsDataAccord_Event() first
}
\examples{
BarrettsBasicNumbers(x,'Endo_ResultPerformed')
}
\keyword{Number}
\keyword{of}
\keyword{therapies}
