% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Event}
\alias{BarrettsDataAccord_Event}
\title{BarrettsDataAccord_Event}
\usage{
BarrettsDataAccord_Event(x, y, z, aa, bb)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The histology main text}

\item{z}{The Procedure Performed column}

\item{aa}{The endoscopic diagnosis column}

\item{bb}{The endoscopic findings column if different to the Diagnosis column}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it extracts the event
Being the procedure done at the time and the follow-up timings
}
\examples{
v<-TheOGDReportFinal
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second Endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
Mypath<-PathDataFrameFinalColon
HistolTree<-list("Hospital Number","Patient Name","DOB:","General Practitioner:",
"Date of procedure:","Clinical Details:","Macroscopic description:","Histology:","Diagnosis:","")
for(i in 1:(length(HistolTree)-1)) {
 Mypath<-Extractor(Mypath,"PathReportWhole",as.character(HistolTree[i]),
 as.character(HistolTree[i+1]),as.character(HistolTree[i]))
}
Mypath$Dateofprocedure<-as.Date(Mypath$Dateofprocedure)
v<-HistolChopperAccessionNumber(Mypath,"Histology","SP-\\\\d{2}-\\\\d{7}")
v<-HistolChopperDx(v,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b<-BarrettsDataAccord_Event(v,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')

}
\keyword{Event}
\keyword{extraction}
