% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Prague}
\alias{BarrettsDataAccord_Prague}
\title{BarrettsDataAccord_Prague}
\usage{
BarrettsDataAccord_Prague(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it extracts the Prague score
}
\examples{
BarrettsDataAccord_Prague(v,'Endo_ResultText')
}
\keyword{Prague}
\keyword{score}
