% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDDRQual}
\alias{BarrettsDDRQual}
\title{Dysplasia detection rate on surveillance}
\usage{
BarrettsDDRQual(dataframe, EndoscopistReportColumn, IMorNoIM)
}
\arguments{
\item{dataframe}{dataframe,}

\item{EndoscopistReportColumn}{column with the Endoscopist names}

\item{IMorNoIM}{extracted with the function Barretts_PathStage()}
}
\description{
This function assesses the dysplasia detection rate per endoscopist.
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolDx(Mypath,'Diagnosis')
v<-HistolExtrapolDx(v,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
#The function relies on the other Barrett's functions being run as well:
b1<-Barretts_PragueScore(v,'Findings')
b2<-Barretts_PathStage(b1,'Histology')
b3<-Barretts_EventType(b2,'Histology',
'ProcedurePerformed','Indications','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-Barretts_FUType(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function takes the column with the extracted worst grade of
# histopathology and returns the proportion of each finding (ie
# proportion with low grade dysplasia, high grade etc.) for each
# endoscopist
jj<-BarrettsDDRQual(b4,'Endoscopist','IMorNoIM')
rm(v)
}
\keyword{detection}
\keyword{dysplasia}
\keyword{rate}
