% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDocumentQual}
\alias{BarrettsDocumentQual}
\title{Analysis of Barrett's Documentation Quality}
\usage{
BarrettsDocumentQual(dataframe, Findings)
}
\arguments{
\item{dataframe}{dataframe}

\item{Findings}{column of interest- usually the main body of the
endoscopic report}
}
\description{
This function assesses the Barrett's documentation. This notes how many
reports contain the mandatory report fields as specified in the BSG standards
 on Barrett's endoscopic reporting. This should be run after the
Barretts_PragueScore.
}
\examples{
# Firstly relevant columns are extrapolated from the
b1<-Barretts_PragueScore(Myendo,'Findings')
# The documentation is really from the endoscopic Findings column
BarrettsDocumentQual(b1,'Findings')
}
\keyword{Documentation}
