% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Barretts.R
\name{BarrettsParisEMR}
\alias{BarrettsParisEMR}
\title{Paris vs histopath Barrett's}
\usage{
BarrettsParisEMR(Column, Column2)
}
\arguments{
\item{Column}{Endoscopy report field of interest as a string vector}

\item{Column2}{Another endoscopy report field of interest as a string vector}
}
\value{
a string vector
}
\description{
This looks at the Paris grades of each EMR and then creates a heatmap
of pathological grade vs
endoscopic Paris grade.This should only be run after all the
BarrettsAll functions.
}
\examples{
# 
Myendo$EMR<-BarrettsParisEMR(Myendo$ProcedurePerformed,Myendo$Findings)
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
