% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPathDetectQual}
\alias{BarrettsPathDetectQual}
\title{Dysplasia detection on surveillance}
\usage{
BarrettsPathDetectQual(dataframe, titlePlot)
}
\arguments{
\item{dataframe}{the dataframe}

\item{titlePlot}{The plot title}
}
\description{
This function assesses pathology of specimens taken at surveillance per year.
It outputs a plot which determines the the overall number of pathologies (
low/high grade dysplasia and cancer) for patients on surveillance
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolDx(Mypath,'Diagnosis')
v<-HistolExtrapolDx(v,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The function relies on the other Barrett's functions being run as well:
b1<-Barretts_PragueScore(v,'Findings')
b2<-Barretts_PathStage(b1,'Histology')
b3<-Barretts_EventType(b2,'Histology',
'ProcedurePerformed','Indications','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-Barretts_FUType(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function simply the the histopathological grades overall for
# your dataset and then creates a frequency plot of them
BarrettsPathDetectQual(b4,'Myplot')
rm(v)
}
\keyword{and}
\keyword{cancer}
\keyword{detection}
\keyword{dysplasia}
