% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPatientTracking_UniqueHospNum}
\alias{BarrettsPatientTracking_UniqueHospNum}
\title{BarrettsPatientTracking_UniqueHospNum}
\usage{
BarrettsPatientTracking_UniqueHospNum(x, rule, PatientID)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{rule}{Rule of interest}

\item{PatientID}{Column containing patient numbers}
}
\description{
This function gets the unique patient ID's for each patient, for each rule. It lists the unique PatientIDs assocaited with a rule ('Rule1','Rule2','Rule3','NoRules')
}
\examples{
BarrettsPatientTracking_UniqueHospNum(b,'Rule1')
}
\keyword{Rule}
