% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsQuality_AnalysisBiopsyNumber}
\alias{BarrettsQuality_AnalysisBiopsyNumber}
\title{BarrettsQuality_AnalysisBiopsyNumber}
\usage{
BarrettsQuality_AnalysisBiopsyNumber(x, Endo_ResultPerformed, PatientID,
  Endoscopist)
}
\arguments{
\item{x}{dataframe}

\item{Endo_ResultPerformed}{Date of the Endocscopy}

\item{PatientID}{Patient's unique identifier}

\item{Endoscopist}{name of the column with the Endoscopist names}
}
\description{
This function gets the biopsies taken per endoscopy and compares to the
Barrett's length
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
BarrettsQuality_AnalysisBiopsyNumber(b,"Endoscopist","Endo_ResultPerformed","Endoscopist")
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
