% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsQuality_AnalysisDocumentation}
\alias{BarrettsQuality_AnalysisDocumentation}
\title{Analysis of Barrett's Documentation Quality}
\usage{
BarrettsQuality_AnalysisDocumentation(x, Findings)
}
\arguments{
\item{x}{dataframe}

\item{Findings}{column of interest- usually the main body of the 
endoscopic report}
}
\description{
This function assesses the Barrett's documentation. This notes how many 
reports contain the mandatory report fields as specified in the BSG standards
 on Barrett's endoscopic reporting. This should be run after the 
BarrettsDataAccord_Prague.
}
\examples{
# Firstly relevant columns are extrapolated from the 
# Mypath demo dataset. These functions are all part of Histology data 
# cleaning as part of the package. 
v<-HistolChopperDx(Mypath,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The pathology stage is then recorded
b<-BarrettsDataAccord_PathStage(v,'Histology')
# The documentation is really from the endoscopic Findings column
BarrettsQuality_AnalysisDocumentation(b,'Findings')
}
\keyword{Documentation}
