% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillance_PathDetection}
\alias{BarrettsSurveillance_PathDetection}
\title{BarrettsSurveillance_PathDetection}
\usage{
BarrettsSurveillance_PathDetection(x, titlePlot)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The plot title}
}
\description{
This function assesses pathology of specimens taken at surveillance per year
}
\examples{
BarrettsSurveillance_PathDetection(x,'Myplot')
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
