% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillanceDDR}
\alias{BarrettsSurveillanceDDR}
\title{BarrettsSurveillanceDDR}
\usage{
BarrettsSurveillanceDDR(x, y, IMorNoIM)
}
\arguments{
\item{x}{dataframe,}

\item{y}{column with the Endoscopist names}

\item{IMorNoIM}{extracted with the function BarrettsDataAccord_PathStage()}
}
\description{
This function assesses the dysplasia detection rate per endoscopist
}
\examples{
v<-TheOGDReportFinal
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
b<-BarrettsDataAccord_PathStage(v,'Histo_ResultText')
BarrettsSurveillanceDDR(b,'Endoscopist','IMorNoIM')
}
\keyword{detection}
\keyword{dysplasia}
\keyword{rate}
