% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapeuticsOnly}
\alias{BarrettsTherapeuticsOnly}
\title{BarrettsTherapeuticsOnly}
\usage{
BarrettsTherapeuticsOnly(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function makes all therapeutics unique
}
\keyword{Therapeutic}
