% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapeuticsRFA_ByCatheter}
\alias{BarrettsTherapeuticsRFA_ByCatheter}
\title{RFA catheter use}
\usage{
BarrettsTherapeuticsRFA_ByCatheter(EndoSubsetRFA, y, z)
}
\arguments{
\item{EndoSubsetRFA}{The dataframe}

\item{y}{report field of interest}

\item{z}{Another endoscopy report field of interest}
}
\description{
This looks at the basic numbers of RFA by catheter type used.
This should only be run after all the BarrettsDataAccord functions.
}
\examples{
v<-HistolChopperDx(Mypath,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,
"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
                             'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsTherapeuticsRFA_ByCatheter(b4,"ProcedurePerformed","Findings")
}
\keyword{RFA,}
\keyword{Radiofrequency}
\keyword{ablation}
