% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapy_Numbers_EMRsByGrade}
\alias{BarrettsTherapy_Numbers_EMRsByGrade}
\title{BarrettsTherapy_Numbers_EMRsByGrade}
\usage{
BarrettsTherapy_Numbers_EMRsByGrade(EndoSubsetEMR)
}
\arguments{
\item{EndoSubsetEMR}{The dataset.}
}
\description{
Plots all the pathological grades of the EMRs.This should only be 
run after all the
BarrettsDataAccord functions.
}
\examples{
v<-HistolChopperDx(Mypath,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsTherapy_Numbers_EMRsByGrade(b4)
}
\keyword{EMR}
\keyword{chart}
