% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{BiopsyIndex}
\alias{BiopsyIndex}
\title{Biopsy index list}
\usage{
BiopsyIndex()
}
\description{
This function returns all the conversions from common version of events to
a standardised event list, much like the Location standardidastion function
This does not include EMR as this is 
extracted from the pathology so is part of pathology type. It is used for
automated OPCS-4 coding.
}
\examples{
# unique(unlist(EventList(), use.names = FALSE))
}
\keyword{Event}
\keyword{extraction}
