% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ChopperNewLines}
\alias{ChopperNewLines}
\title{EndoscChopperNewLines}
\usage{
ChopperNewLines(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{The endoscopy text column}
}
\description{
Cleans the long lines to separate them into new lines if necessary
}
\examples{
ChopperNewLines(NewData,'Endo_ResultText')
}
\keyword{Endoscopy}
\keyword{Newlines}
