% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{ColonEndoRaw}
\alias{ColonEndoRaw}
\title{ColonEndoRaw}
\usage{
ColonEndoRaw(x)
}
\arguments{
\item{x}{None needed}
}
\description{
Creates raw Pathology reports
}
\examples{
ColonEndoRaw(x)
}
\keyword{Pathology}
\keyword{reports}
