% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ColonFinal}
\alias{ColonFinal}
\title{Fake Lower GI Endoscopy Set}
\format{
A data frame with 2000 rows and 1 variables:
\describe{
  \item{OGDReportWhole}{The whole report, in text}
}
}
\usage{
ColonFinal
}
\description{
A dataset containing fake lower GI endoscopy reports. The report field is provided as a whole report
without any fields having been already extracted
}
\keyword{datasets}
