% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{ColonpathRep}
\alias{ColonpathRep}
\title{ColonpathRep}
\usage{
ColonpathRep(x)
}
\arguments{
\item{x}{None needed}
}
\description{
Creates raw Pathology reports
}
\examples{
ColonpathRep(x)
}
\keyword{Pathology}
\keyword{reports}
