% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{ColumnCleanUp}
\alias{ColumnCleanUp}
\title{Tidy up messy columns}
\usage{
ColumnCleanUp(vector)
}
\arguments{
\item{vector}{column of interest}
}
\value{
This returns a character vector
}
\description{
This does a general clean up of whitespace,
semi-colons,full stops at the start
of lines and converts end sentence full stops to new lines.
}
\examples{
ii<-ColumnCleanUp(Myendo$Findings)
}
\keyword{Cleaner}
