% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ColumnCleanUpAll}
\alias{ColumnCleanUpAll}
\title{Tidies up all columns}
\usage{
ColumnCleanUpAll(dataframe)
}
\arguments{
\item{dataframe}{dataframe with column of interest}
}
\description{
This does a general clean up of whitespace,
semi-colons,full stops at the start
of lines and converts end sentence full stops to new lines.
It it used as part of the Extractor function
It can be used as part of the other functions
or as a way of providing a 'positive diagnosis only' type output (see
HistolChopperDx)
}
\examples{
Myendo<-lapply(Myendo, ColumnCleanUpAll)

}
\keyword{Cleaner}
