% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{dev_ExtrapolateEndoscopicICD10}
\alias{dev_ExtrapolateEndoscopicICD10}
\title{Primary Diagnosis from Endoscopy}
\usage{
dev_ExtrapolateEndoscopicICD10(dataframe, Procedure, PathSite, FINDINGS,
  ENDOSCOPICDIAGNOSIS, EndoscopyEvent)
}
\arguments{
\item{dataframe}{the dataframe}

\item{Procedure}{the Procedure column}

\item{PathSite}{the PathSite column}

\item{FINDINGS}{the FINDINGS column}

\item{ENDOSCOPICDIAGNOSIS}{the ENDOSCOPICDIAGNOSIS column}

\item{EndoscopyEvent}{the EndoscopyEvent column}
}
\description{
This function extracts the primary diagnosis from the endoscopy free text. It will also add
pathology diagnosis to the final primary code
}
\examples{
# Need to run the HistolTypeSite and EndoscopyEvent functions first here
# SelfOGD_Dunn$OPCS4w<-ExtrapolateOPCS4Prep(SelfOGD_Dunn,"PROCEDUREPERFORMED",
# "PathSite","EndoscopyEvent")
}
\keyword{OPCS-4}
\keyword{codes}
\keyword{extraction}
