% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{dev_ExtrapolateOPCS4Prep}
\alias{dev_ExtrapolateOPCS4Prep}
\title{OPCS-4 Coding}
\usage{
dev_ExtrapolateOPCS4Prep(dataframe, Procedure, PathSite, Event, extentofexam)
}
\arguments{
\item{dataframe}{the dataframe}

\item{Procedure}{The Procedure column}

\item{PathSite}{The column containing the Pathology site}

\item{Event}{the EVENT column}

\item{extentofexam}{the furthest point reached in the examination}
}
\description{
This function extracts the OPCS-4 codes for all Barrett's procedures
It should take the OPCS-4 from the EVENT and perhaps also using extent
depending on how the coding is done. The EVENT column will need to 
extract multiple findings
The hope is that the OPCS-4 column will then map from the EVENT column. This returns a nested list 
column with the procedure, furthest path site and event performed
}
\examples{
# Need to run the HistolTypeSite and EndoscopyEvent functions first here
# SelfOGD_Dunn$OPCS4w<-ExtrapolateOPCS4Prep(SelfOGD_Dunn,"PROCEDUREPERFORMED",
# "PathSite","EndoscopyEvent")
}
\keyword{OPCS-4}
\keyword{codes}
\keyword{extraction}
