% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{dev_SensAndSpecif}
\alias{dev_SensAndSpecif}
\title{Sensitivity and Specificity of text mining functions}
\usage{
dev_SensAndSpecif(dataframe, ref, actual)
}
\arguments{
\item{dataframe}{dataframe}

\item{ref}{The column with the reference values}

\item{actual}{The column with the actual values}
}
\description{
This is part of the validation workflow. The idea is that 
There is a dataframe which contains a column with the reference value
and another column with the actual value. If the test or reference has not
detected a value then this is a negative and is labelled as 'Insufficient'
}
\examples{
#To be set up 
}
\keyword{and}
\keyword{sensitivity}
\keyword{specificity}
