% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{DictionaryInPlaceReplace}
\alias{DictionaryInPlaceReplace}
\title{Dictionary In Place Replace}
\usage{
DictionaryInPlaceReplace(inputString, list)
}
\arguments{
\item{inputString}{the input string}

\item{list}{The replacing list}
}
\value{
This returns a character vector
}
\description{
This maps terms according to the lexicons and replaces them with the 
standardised term within the text. It is used within the textPrep function
}
\examples{
inputText<-DictionaryInPlaceReplace(TheOGDReportFinal$OGDReportWhole,LocationList())
}
\keyword{Replace}
