% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\docType{package}
\name{EbdoMineR}
\alias{EbdoMineR}
\alias{EbdoMineR-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

