% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoBasicGraph}
\alias{EndoBasicGraph}
\title{Basic graph creation using the template specified in theme_Publication.}
\usage{
EndoBasicGraph(dataframe, xdata, number)
}
\arguments{
\item{dataframe}{dataframe}

\item{xdata}{The x column}

\item{number}{The numeric column}
}
\value{
Myplot This is the final plot

Myplot
}
\description{
This creates a basic graph using the template specified in theme_Publication.
It takes a numeric column and plots it against any non-numeric x axis in a ggplot
}
\examples{
# This function plots numeric y vs non-numeric x
# Get some numeric columns e.g. number of biopsies and size
Mypath$Size <- HistolBxSize(Mypath$Macroscopicdescription)
Mypath$NumBx <- HistolNumbOfBx(Mypath$Macroscopicdescription, "specimen")
Mypath2 <- Mypath[, c("NumBx", "Size")]
EndoBasicGraph(Mypath, "Size", "NumBx")
}
\seealso{
Other Data Presentation helpers: \code{\link{scale_colour_Publication}},
  \code{\link{scale_fill_Publication}},
  \code{\link{theme_Publication}}
}
\concept{Data Presentation helpers}
\keyword{Time}
\keyword{plots}
