% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoClean}
\alias{EndoClean}
\title{Cleans up the endoscopy columns}
\usage{
EndoClean(dataframe, ...)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{Column}{column of interest}
}
\description{
This function runs all of the cleaning subfunctions rather than needing to 
run them individually
}
\examples{
v<-EndoClean(Myendo,Findings='Findings2')
}
\keyword{Clean}
