% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoDataVizEDA}
\alias{EndoDataVizEDA}
\title{Create a Circos plot for patient flow}
\usage{
EndoDataVizEDA(dataframe, Title)
}
\arguments{
\item{dataframe}{dataframe}

\item{Title}{The plot title}
}
\value{
Myplot the EDA final plot
}
\description{
This allows us to look at the overall flow from one
type of procedure to another using circos plots.
}
\examples{
# This function does EDA
}
\keyword{Circos}
