% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoDataVizNum_xNum_y}
\alias{EndoDataVizNum_xNum_y}
\title{This creates a graph with any categorical data along the x-axis
and numeric data along teh y-axis. It implements the template styles as per 
theme_Publication.}
\usage{
EndoDataVizNum_xNum_y(dataframe, xdata, number)
}
\arguments{
\item{dataframe}{dataframe}

\item{xdata}{The Time column}

\item{number}{The numeric column}
}
\value{
Myplot the EDA final plot

Myplot
}
\description{
This creates a graph with any categorical data along the x-axis
and numeric data along teh y-axis. It implements the template styles as per 
theme_Publication.
}
\examples{
# This function plot numeric y vs non-numeric x
#Get some numeric columns eg number of biopsies and size
Mypath$Size<-HistolBxSize(Mypath$Macroscopicdescription)
Mypath$NumBx<-HistolNumbOfBx(Mypath$Macroscopicdescription,'specimen')
Mypath2<-Mypath[,c("NumBx","Size")]
EndoDataVizNum_xNum_y(Mypath,'Size','NumBx')
}
\keyword{Time}
\keyword{plots}
