% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoDataVizNum_xProp_y}
\alias{EndoDataVizNum_xProp_y}
\title{Create non Numeric x versus y}
\usage{
EndoDataVizNum_xProp_y(dataframe, b, Proportion)
}
\arguments{
\item{dataframe}{dataframe}

\item{b}{The numeric column}

\item{Proportion}{The Proportion column}
}
\value{
Myplot
}
\description{
This allows us to look at the overall flow from one
type of procedure to another using circos plots.
}
\examples{
# This function plot numeric y vs non-numeric x


}
\keyword{plot}
\keyword{proportions}
