% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoDataVizTime_xNum_y}
\alias{EndoDataVizTime_xNum_y}
\title{Create non Numeric x versus y}
\usage{
EndoDataVizTime_xNum_y(dataframe, xdata, number)
}
\arguments{
\item{dataframe}{dataframe}

\item{xdata}{The Time column}

\item{number}{The numeric column}

\item{HospNum_Id}{Column with the patient's unique hospital number}
}
\value{
Myplot the EDA final plot

Myplot
}
\description{
This allows us to look at the overall flow from one
type of procedure to another using circos plots.
}
\examples{
# This function plot numeric y vs non-numeric x


}
\keyword{Time}
\keyword{plots}
