% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{EndoPaste}
\alias{EndoPaste}
\title{Paste endoscopy and histology results into one}
\usage{
EndoPaste(x)
}
\arguments{
\item{x}{the dataframe}
}
\value{
This returns a list with a dataframe containing one column of the merged text
and a character vector which is the delimiter list for when the textPrep function is used
}
\description{
As spreadsheets are likely to be submitted with pre-segregated data as appears from 
endoscopy software output, these should be remerged prior to cleaning. This function
takes the column headers and places it before each text so that the original
full text is recreated. It will use the column headers as the delimiter. This should 
be used before textPrep as the textPrep function takes a character vector (ie the whole
report and not a segregated one) only
}
\examples{
testList<-structure(list(PatientName = c("Tom Hardy", "Elma Fudd", "Bingo Man"
), HospitalNumber = c("H55435", "Y3425345", "Z343424"), Text = c("All bad. Not good", 
"Serious issues", "from a land far away")), class = "data.frame", row.names = c(NA, -3L))
EndoPaste(testList)
}
\concept{NLP - Text merging:}
\keyword{Merge}
\keyword{columns}
\keyword{dataframe}
\keyword{into}
\keyword{one}
\keyword{text}
