% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoPOS}
\alias{EndoPOS}
\title{Parts of speech tagging for reports}
\usage{
EndoPOS(dataframe)
}
\arguments{
\item{dataframe}{dataframe}

\item{MacroColumn}{column to extract the numbers from. Usually the column
with the Nature of the specimen or the Macroscopic description in it}
}
\description{
This uses udpipe to tag the text. It then compresses all of the text 
so you have continuous POS tagging or the whole text
}
\examples{
pp<-HistolMacDescrip(Mypath$Macroscopicdescription)
}
\keyword{Macroscopic}
