% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{EndoRaw2}
\alias{EndoRaw2}
\title{EndoRaw}
\usage{
EndoRaw2()
}
\arguments{
\item{x}{None needed}
}
\description{
Generates fake Endoscopy date
}
\examples{
EndoRaw(x)
}
\keyword{data}
\keyword{endoscopy}
\keyword{fake}
