% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscAll}
\alias{EndoscAll}
\title{Extracts the columns from the raw report}
\usage{
EndoscAll(dataframe)
}
\arguments{
\item{dataframe}{the dataframe}
}
\description{
This is the parent cleaning function for the endoscopy report. It contains
all the other functions for the endoscopy report to be cleaned up. It
relies on the columns being named in a standardised way as below
}
\examples{
# Rename the columns in whatever endoscopy dataframe you have
names(Myendo)<-c("OGDReportWhole","HospitalNumber","PatientName",
"GeneralPractitioner","Dateofprocedure","Endoscopist","Secondendoscopist",
"Medications","Instrument","ExtentofExam","Indications","ProcedurePerformed",
"Findings" )
#Now use the function
EndoscAll(Myendo)
rm(Myendo)
}
\keyword{Extraction}
