% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperEndoscopist}
\alias{EndoscChopperEndoscopist}
\title{EndoscChopperEndoscopist}
\usage{
EndoscChopperEndoscopist(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{The endoscopy text column}
}
\description{
This extracts the endoscopist from the report
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v$Endo_ResultPerformed<-as.Date(v$Endo_ResultPerformed,format="\%d/\%m/\%Y")
v<-EndoscChopperEndoscopist(v,"Endoscopist")
}
\keyword{Endoscopist}
\keyword{extraction}
