% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperIndications}
\alias{EndoscChopperIndications}
\title{Cleans indications column if present}
\usage{
EndoscChopperIndications(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the Indication from the report assuming such a column exists. 
It largely gets rid of carriage returns especially if used after the 
ChopperNewLinesfunction. There may be multiple indications.
It should be used after the Extractor and the optional ChopperNewLines has 
been used.
}
\examples{
v<-EndoscChopperIndications(Myendo,'Indications')
}
\keyword{Indications}
