% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperInstrument}
\alias{EndoscChopperInstrument}
\title{EndoscChopperInstrument}
\usage{
EndoscChopperInstrument(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the Instrument from the report
}
\examples{
v<-EndoscChopperInstrument(Myendo,'Instrument')
}
\keyword{Instrument}
