% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperMeds}
\alias{EndoscChopperMeds}
\title{EndoscChopperMeds}
\usage{
EndoscChopperMeds(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the medications from the report
}
\examples{
v<-EndoscChopperMeds(Myendo,'Medications')
}
\keyword{Endoscopy}
\keyword{medications}
