% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperProcPerformed}
\alias{EndoscChopperProcPerformed}
\title{EndoscChopperProcPerformed}
\usage{
EndoscChopperProcPerformed(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the procedure performed from the report
}
\examples{
v<-EndoscChopperProcPerformed(Myendo,'ProcedurePerformed')
}
\keyword{Procedure}
