% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscFindings}
\alias{EndoscFindings}
\title{Cleans Procedure performed column if present}
\usage{
EndoscFindings(dataframe, FindingsColumn)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{FindingsColumn}{column of interest}
}
\description{
This cleans the Findings column from the report assuming
such a column exists. Findings relates to what was found at the endoscopy
This is usually a separate entry to the overall 'Diagnosis' but any
are in which the description of the endoscopic findings, including
overall diagnosis or not, can be used.
It should be used after the Extractor and the optional NewLines
has been used. At present it only cleans cm measurement
}
\examples{
gg<-EndoscFindings(Myendo,'Findings')
}
\keyword{Procedure}
