% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscIndications}
\alias{EndoscIndications}
\title{Cleans indications column if present}
\usage{
EndoscIndications(dataframe, IndicationColumn)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{IndicationColumn}{column of interest}
}
\description{
This cleans the Indication from the report assuming such a column exists.
It largely gets rid of carriage returns especially if used after the
NewLinesfunction. There may be multiple indications.
It should be used after the Extractor and the optional NewLines has
been used.
}
\examples{
v<-EndoscIndications(Myendo,'Indications')
}
\keyword{Indications}
