% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{EndoscopicICD10}
\alias{EndoscopicICD10}
\title{Primary Diagnosis from Endoscopy}
\usage{
EndoscopicICD10(dataframe, Procedure, PathSite, FINDINGS,
  ENDOSCOPICDIAGNOSIS, EndoscopyEvent)
}
\arguments{
\item{dataframe}{the dataframe}

\item{EVENT}{the EVENT column}
}
\description{
This function extracts the primary diagnosis from the endoscopy free text. It will also add
pathology diagnosis to the final primary code
}
\examples{
# Need to run the HistolTypeSite and EndoscopyEvent functions first here
SelfOGD_Dunn$OPCS4w<-OPCS4Prep(SelfOGD_Dunn,"PROCEDUREPERFORMED","PathSite","EndoscopyEvent")
}
\keyword{OPCS-4}
\keyword{codes}
\keyword{extraction}
