% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Endoscopies}
\alias{Endoscopies}
\title{Fake Endoscopy data set}
\format{A data frame with 1000 rows and 10 variables:
\describe{
  \item{HospNum_Id}{Unique Hospintal Number ID, in text}
  \item{EndoReports}{The whole report, in text}
  \item{Date}{Date the endoscopy was done, as a date}
  \item{Endoscopist}{Name of the endoscopist, in text}
  \item{Midazolam}{Number of mg of midazolam given, in numeric}
  \item{Fentanyl}{Number of mcg of Fentanyl given, in numeric}
  \item{Indication}{Indication for the procedure, in text}
  \item{Diagnosis}{The written report's diagnosis, in text}
  \item{BarrC}{Prague circumferential length, in cm}
  \item{BarrM}{Pracue maximal length, in cm}
}}
\usage{
Endoscopies
}
\description{
A dataset containing fake endoscopy reports. The report fields have already been
separated into their own fields as opposed to TheOGDReportFinal and ColonFinal
}
\keyword{datasets}
