% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{EndoscopyEvent}
\alias{EndoscopyEvent}
\title{Extract the endoscopic event.}
\usage{
EndoscopyEvent(dataframe, EventColumn1, Procedure, Macroscopic, Histology)
}
\arguments{
\item{dataframe}{datafrane of interest}

\item{EventColumn1}{The relevant endoscopt free text column describing the findings}

\item{Procedure}{Column saying which procedure was performed}

\item{Macroscopic}{Column describing all the macroscopic specimens}

\item{Histology}{Column with free text histology (usually microscopic histology)}
}
\value{
This returns a character vector
}
\description{
This extracts the endoscopic event. It looks for the event 
term and then looks in the event sentence as well as the one above to see if
the location is listed. It only looks within the endoscopy fields. If tissue is taken
then this will be extracted with the HistolTypeAndSite function rather than being 
listed as a result as this is cleaner and more robust.
}
\examples{
# Myendo$EndoscopyEvent<-EndoscopyEvent(Myendo,"Findings",
# "ProcedurePerformed","MACROSCOPICALDESCRIPTION","HISTOLOGY")
}
\seealso{
Other Endoscopy specific cleaning functions: \code{\link{EndoscEndoscopist}},
  \code{\link{EndoscInstrument}}, \code{\link{EndoscMeds}}
}
\concept{Endoscopy specific cleaning functions}
\keyword{Find}
\keyword{and}
\keyword{replace}
