% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EntityPairs_OneSentence}
\alias{EntityPairs_OneSentence}
\title{EntityPairs_OneSentence}
\usage{
EntityPairs_OneSentence(inputText, list1, list2)
}
\arguments{
\item{EventColumn1}{The relevant pathology text column}

\item{EventColumn2}{The alternative pathology text column}
}
\description{
Use to see if words from two lists co-exist within a sentence. Eg site and tissue type.
}
\examples{
# tbb<-EntityPairs_OneSentence(Mypath$Histology,HistolType(),LocationList())
}
\keyword{PathPairLookup}
