% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EntityPairs_TwoSentence}
\alias{EntityPairs_TwoSentence}
\title{EntityPairs_TwoSentence}
\usage{
EntityPairs_TwoSentence(inputString, list1, list2)
}
\arguments{
\item{inputString}{The relevant pathology text column}

\item{list1}{The intial list to assess}

\item{list2}{The other list to look for}
}
\description{
This is used to look for relationships between site and event especially for endoscopy events
}
\examples{
# tbb<-EntityPairs_TwoSentence(Myendo,"Findings")
}
\keyword{Find}
\keyword{and}
\keyword{replace}
