% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{EventList}
\alias{EventList}
\title{Event list}
\usage{
EventList()
}
\description{
<<<<<<< HEAD
This function returns all the conversions from common version of events to
a standardised event list, much like the Location standardisation function
This does not include EMR as this is
=======
This function returns all the conversions from common version of events to 
a standardised event list, much like the Location standardisation function.
This does not include EMR as this is 
>>>>>>> master
extracted from the pathology so is part of pathology type.
}
\examples{
# unique(unlist(EventList(), use.names = FALSE))
}
\keyword{Event}
\keyword{extraction}
