% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{Extractor}
\alias{Extractor}
\title{Extracts the columns from the raw report}
\usage{
Extractor(x, y, stra, strb, t)
}
\arguments{
\item{x}{the dataframe}

\item{y}{the column to extract from}

\item{stra}{the start of the boundary to extract}

\item{strb}{the end of the boundary to extract}

\item{t}{the column name to create}
}
\description{
This is the main extractor for the Endoscopy and Histology report.
This depends on the user creating a list of words or characters that
act as the words that should be split against. The list is then fed to the
Extractor in a loop so that it acts as the beginning and the end of the
regex used to split the text. Whatever has been specified in the list
is used as a column header. Column headers don't tolerate special characters
like : or ? and / and don't allow numbers as the start character so these
have to be dealt with in the text before processing
}
\examples{
# As column names cant start with a number, one of the dividing
# words has to be converted
Myendo$OGDReportWhole<-gsub('2nd Endoscopist:','Second endoscopist:',
Myendo$OGDReportWhole)
# A list of dividing words (which will also act as column names)
# is then constructed
EndoscTree<-list('Hospital Number:','Patient Name:','General Practitioner:',
'Date of procedure:','Endoscopist:','Second Endoscopist:','Medications',
'Instrument','Extent of Exam:','Indications:','Procedure Performed:',
'Findings:','Endoscopic Diagnosis:')
# The Extractor function is then used as part of a loop to divide the raw
# report up according to the user-defined words in the extraction list
# defined above
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,'OGDReportWhole',as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
res<-Myendo
}
\keyword{Extraction}
