% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ExtrapolatefromDictionary}
\alias{ExtrapolatefromDictionary}
\title{Extrapolate from Dictionary}
\usage{
ExtrapolatefromDictionary(inputString, list)
}
\arguments{
\item{dataframe}{The dataframe}

\item{SampleLocation}{Column describing the Macroscopic sample from histology}
}
\description{
Provides term mapping and extraction in one
Standardises any term according to a mapping lexicon provided and then
extracts the term. Thus is
different to the DictionaryInPlaceReplace in that it provides a new column
with the extracted terms as opposed to changing it in place
}
\examples{
#Firstly we extract histology from the raw report
# using the extractor function
mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
MypathExtraction<-Extractor(PathDataFrameFinal,"PathReportWhole",mywords)
names(MypathExtraction)[names(MypathExtraction) == 'Datereceived'] <- 'Dateofprocedure'
MypathExtraction$Dateofprocedure <- as.Date(MypathExtraction$Dateofprocedure)
# The function then standardises the histology terms through a series of
# regular expressions
ll<-ExtrapolatefromDictionary(Mypath$Histology,HistolType())
rm(MypathExtraction)
}
\keyword{Withdrawal}
