% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ExtrapolateFromLists}
\alias{ExtrapolateFromLists}
\title{Extrapolation from lists}
\usage{
ExtrapolateFromLists(Column, mylist)
}
\arguments{
\item{Column}{The column you want to extract from}

\item{mylist}{The list of values}

\item{dataframe}{dataframe}
}
\description{
This takes a list of terms that you want to look for and
then extracts these into a new column
}
\examples{
#To be set up 
}
\keyword{and}
\keyword{sensitivity}
\keyword{specificity}
