% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{GISymptomsList}
\alias{GISymptomsList}
\title{Index of GI symptoms}
\usage{
GISymptomsList()
}
\description{
This function returns all the common GI symptoms. 
They are simply listed as is without grouping
or mapping. They have been derived from a manual 
list with synonyms derived from the UMLS Methatharus
using the browser.
}
\seealso{
Other NLP - Lexicons: 
\code{\link{BiopsyIndex}()},
\code{\link{EventList}()},
\code{\link{HistolType}()},
\code{\link{LocationListLower}()},
\code{\link{LocationListUniversal}()},
\code{\link{LocationListUpper}()},
\code{\link{LocationList}()},
\code{\link{RFACath}()},
\code{\link{WordsToNumbers}()}
}
\concept{NLP - Lexicons}
\keyword{Event}
\keyword{extraction}
