% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{GRS_Type_Assess_By_Unit}
\alias{GRS_Type_Assess_By_Unit}
\title{GRS_Type_Assess_By_Unit
This extracts the polyps types from the data(for colonoscopy and flexible sigmoidosscopy data)}
\usage{
GRS_Type_Assess_By_Unit(x, ProcPerformed, Endo_Endoscopist, Dx, Histol)
}
\arguments{
\item{x}{The dataframe}

\item{ProcPerformed}{The column containing the Procedure type performed}

\item{Endo_Endoscopist}{column containing the Endoscopist name}

\item{Dx}{The column with the Histological diagnosis}

\item{Histol}{The column with the Histology text in it}
}
\description{
GRS_Type_Assess_By_Unit
This extracts the polyps types from the data(for colonoscopy and flexible sigmoidosscopy data)
}
\examples{
v<-MyColonpath
Histoltree<-list('Clinical Details','Nature of specimen','Histology','Diagnosis','')
v<-ChopperNewLines(v,'Histo_ResultText')
for(i in 1:(length(Histoltree)-1)) {
 v<-Extractor(v,'Histo_ResultText',as.character(Histoltree[i]),
 as.character(Histoltree[i+1]),gsub(' ','',as.character(Histoltree[i])))
}
GRSTable<-GRS_Type_Assess_By_Unit(v,'ProcedurePerformed', 'Endoscopist','Diagnosis','Histology')
}
\keyword{Withdrawal}
