% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolAccessionNumber}
\alias{HistolAccessionNumber}
\title{#' Extract the histology data from the report by removing negative findings
#'
#' This extracts Histology details data from the report and also removes
#' negative findings. The Histology details
#' usually relate to the description of the histological report.
#' @param dataframe dataframe with column of interest
#' @param HistolColumn column of interest
#' @keywords Histology
#' @export
#' @importFrom stringr str_replace
#' @examples ll<-HistolHistol(Mypath,'Histology')}
\usage{
HistolAccessionNumber(dataframe, AccessionColumn, regString)
}
\arguments{
\item{dataframe}{dataframe name and}

\item{AccessionColumn}{the column name as a string.}

\item{regString}{regular expression needed as a string}
}
\description{
HistolHistol <- function(dataframe, HistolColumn) {
  dataframe<-data.frame(dataframe)
  # HISTOLOGY
  dataframe[, HistolColumn] <- str_replace(dataframe[, HistolColumn],
                                           "\n|\r", " ")
  dataframe[, HistolColumn] <- ColumnCleanUp(dataframe, HistolColumn)
  dataframe[, HistolColumn] <- gsub("- ", "\n", dataframe[, HistolColumn],
                                      fixed = TRUE)
  dataframe[, HistolColumn] <- gsub("-[A-Z]", "\n",
                                      dataframe[, HistolColumn]
                                      , fixed = TRUE)
  dataframe$Histol_Simplified<- NegativeRemove(dataframe, HistolColumn)
  dataframe$Histol_Simplified <-str_replace(dataframe$Histol_Simplified,":", "")
  return(dataframe)
}
Extract histological accession number
}
\details{
This extracts Accession Number data data from the report where one is
present. The Accession number relates to the actual specimen number as
ascribed by the pathology service.
}
\examples{
mm<-HistolAccessionNumber(Mypath,'Histology',
"SP-\\\\d{2}-\\\\d{7}")
}
\keyword{Accession}
\keyword{Sample}
\keyword{number}
