% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolBiopsyIndex}
\alias{HistolBiopsyIndex}
\title{HistolBiopsyIndex}
\usage{
HistolBiopsyIndex(pathSiteString)
}
\arguments{
\item{pathSiteString}{The column that has the pathology locatio and tissue type from HistolTypeAndSite}
}
\description{
This returns a number for all the biopsies taken based on distance from orifice. It is for biopsies only
YOU COULD DITCH THIS ONE AS IT IS JUCT A FUNCTION TO EXTRPOLATE FROM A DICTIONARY 
AS PER THE FUNCTION INSIDE THIS FUNCTION
}
\examples{
# SelfOGD_Dunn<-read_excel("/home/rstudio/GenDev/DevFiles/EndoMineRFunctionDev/SelfOGD_Dunn.xlsx")
SelfOGD_Dunn$PathSite<-HistolTypeAndSite(SelfOGD_Dunn,"MACROSCOPICALDESCRIPTION","HISTOLOGY")
HistolBiopsyIndex(SelfOGD_Dunn$PathSite) 
}
\keyword{Pathology}
\keyword{biopsy}
\keyword{index}
