% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperAccessionNumber}
\alias{HistolChopperAccessionNumber}
\title{HistolChopperAccessionNumber}
\usage{
HistolChopperAccessionNumber(x, y, stra)
}
\arguments{
\item{x}{the dataframe name and}

\item{y}{the column name as a string.}

\item{stra}{regular expression needed as a string}
}
\description{
This extracts Accession Number data data from the report
}
\examples{
Mypath<-PathDataFrameFinalColon
Histoltree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Clinical Details","Nature of specimen",
"Histology","Diagnosis","")
for(i in 1:(length(Histoltree)-1)) {
  Mypath<-Extractor(Mypath,"PathReportWhole",as.character(Histoltree[i]),
  as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
v<-HistolChopperAccessionNumber(Mypath,"Histology","SP-\\\\d{2}-\\\\d{7}")
}
\keyword{Accession}
\keyword{Sample}
\keyword{number}
