% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperAll}
\alias{HistolChopperAll}
\title{Clean up all Histology data}
\usage{
HistolChopperAll(dataframe)
}
\arguments{
\item{dataframe}{dataframe with column of interest}
}
\description{
This is a parent function for all the functions below
}
\examples{
names(Mypath)<-c("HospitalNumber","PatientName","DOB",
"GeneralPractitioner","Datereceived","ClinicalDetails",
"Macroscopicdescription","Histology","Diagnosis","HospitalNumber",
"PatientName","DOB","GeneralPractitioner","Dateofprocedure",
"ClinicalDetails","Macroscopicdescription","Histology","Diagnosis")
v<-HistolChopperAll(Mypath)
}
\keyword{Macroscopic}
