% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperExtrapolDx}
\alias{HistolChopperExtrapolDx}
\title{HistolChopperExtrapolDx}
\usage{
HistolChopperExtrapolDx(x, y)
}
\arguments{
\item{x}{the dataframe containing histology results,}

\item{y}{the column to extract dysplasia, cancer, and GIST from- often the Histology diagnosis column}
}
\description{
This extracts other specific diagnoses from the report
}
\examples{
v<-Endomerge(MyColonpath,MyColonendo)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
Histoltree<-list("Clinical Details","Nature of specimen","Histology","Diagnosis","")
v<-ChopperNewLines(v,"Histo_ResultText")
for(i in 1:(length(Histoltree)-1)) {
 v<-Extractor(v,"Histo_ResultText",as.character(Histoltree[i]),as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
v<-HistolChopperExtrapolDx(v,"Diagnosis")
}
\keyword{Histology}
\keyword{diagnosis}
