% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperHistol}
\alias{HistolChopperHistol}
\title{HistolChopperHistol}
\usage{
HistolChopperHistol(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This extracts Histology details data from the report
}
\examples{
I often use '[Mm][Ii][Cc][Rr][oo][Ss][Cc][Oo][Pp][Ii][Cc]|HISTOLOGY' 
and '[Dd][Ii][Aa][Gg][Nn][Oo][Ss][Ii][Ss]'
}
\keyword{Histology}
