% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescrip}
\alias{HistolChopperMacDescrip}
\title{HistolChopperMacDescrip}
\usage{
HistolChopperMacDescrip(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{column to extract the numbers from. Usually the column
with the Nature of the specimen or the Macroscopic description in it}
}
\description{
This extracts numbers from written (spelt) numbers
}
\examples{
HistolChopperMacDescrip(Mypath, "Macroscopicdescription")
}
\keyword{Macroscopic}
