% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescripCleanup}
\alias{HistolChopperMacDescripCleanup}
\title{HistolChopperMacDescripCleanup}
\usage{
HistolChopperMacDescripCleanup(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest that describes the macroscopic specimen}
}
\description{
This extracts Macroscopic description data from the report
}
\examples{
HistolChopperHistol(Mypath,'Histology')
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperMacDescripCleanup(v,"Macroscopicdescription")
}
\keyword{Macroscopic}
