% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperNumbOfBx}
\alias{HistolChopperNumbOfBx}
\title{Extract the number of biopsies taken from histology report}
\usage{
HistolChopperNumbOfBx(dataframe, MacroColumn, regString)
}
\arguments{
\item{dataframe}{the dataframe}

\item{MacroColumn}{Column containing the Macroscopic description text}

\item{regString}{The keyword to remove and to stop at in the regex}
}
\description{
This extracts the number of biopsies taken from the pathology report.
This is usually from the Macroscopic description column.
It collects everything from the regex [0-9]{1,2}.{0,3}
to whatever the string boundary is (z).
}
\examples{
v<-HistolChopperNumbOfBx(Mypath,'Macroscopicdescription',
'specimen')
}
\keyword{Biopsy}
\keyword{number}
