% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolDx}
\alias{HistolDx}
\title{Extracts histological diagnosis}
\usage{
HistolDx(dataframe, HistolColumn)
}
\arguments{
\item{dataframe}{dataframe}

\item{HistolColumn}{column containing the Hisopathology report}
}
\description{
This extracts Diagnosis data from the report. The Diagnosis is the overall
impression of the pathologist for that specimen. At the moment, Only Capital
D included (not lower case d) to make sure picks up subtitle header as
opposed to mentioning 'diagnosis' as part of a sentence.  Column specific
cleanup and negative remover have also been implemented here.
}
\examples{
v<-HistolDx(Mypath,'Diagnosis')
}
\keyword{Diagnosis}
\keyword{Histology}
