% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolHistol}
\alias{HistolHistol}
\title{Extract the histology data from the report by removing negative findings}
\usage{
HistolHistol(dataframe, HistolColumn)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{HistolColumn}{column of interest}
}
\description{
This extracts Histology details data from the report. The Histology details
usually relate to the description of the histological report. This
may be refined in further iterations.
}
\examples{
t<-HistolHistol(Mypath,'Histology')
}
\keyword{Histology}
