% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolMacDescripCleanup}
\alias{HistolMacDescripCleanup}
\title{Clean up histological macroscopic description data}
\usage{
HistolMacDescripCleanup(dataframe, MacroColumn)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{MacroColumn}{the dataframe column with the Macroscopic description text}
}
\description{
This is the parent cleaning function for the histology report. It contains
all the other functions for the histology report to be cleaned up. It
relies on the columns being named in a standardised way as below:
}
\examples{
 

mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
Mypath<-Extractor(Mypath,"PathReportWhole",mywords)
names(Mypath)<-c("HospitalNumber","PatientName","DOB",
"GeneralPractitioner","Datereceived","ClinicalDetails",
"Macroscopicdescription","Histology","Diagnosis","HospitalNumber",
"PatientName","DOB","GeneralPractitioner","Dateofprocedure",
"ClinicalDetails","Macroscopicdescription","Histology","Diagnosis")
HistolMacDescripCleanup(Mypath,"Histology")
rm(Mypath)
}
\keyword{Histology}
