% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{HistolType}
\alias{HistolType}
\title{Extract pathology type}
\usage{
HistolType()
}
\arguments{
\item{dataframe}{dataframe with column of interest}
}
\description{
This standardizes terms to describe the pathology tissue type being exmained
}
\examples{
#No examples as just returns a list
}
\keyword{Pathology}
\keyword{type}
