% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolTypeAndSite}
\alias{HistolTypeAndSite}
\title{HistolTypeAndSite}
\usage{
HistolTypeAndSite(dataframe, Procedure, EventColumn1, EventColumn2)
}
\arguments{
\item{Procedure}{The procedure performed}

\item{EventColumn1}{The relevant pathology text column}

\item{EventColumn2}{The alternative pathology text column}
}
\description{
This needs some blurb to be written. Used in the OPCS4 coding
}
\examples{
# SelfOGD_Dunn$PathSite<-HistolTypeAndSite(SelfOGD_Dunn,"PROCEDUREPERFORMED","MACROSCOPICALDESCRIPTION","HISTOLOGY")
}
\keyword{Find}
\keyword{and}
\keyword{replace}
