% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Histop_df}
\alias{Histop_df}
\title{Fake Colon Endoscopy Set}
\format{A data frame with 2000 rows and 1 variables:
\describe{
  \item{HospNum_Id}{Unique Hospintal Number ID, in text}
  \item{HistoReport}{The whole report, in text}
  \item{Date}{Date the endoscopy was done, as a date}
  \item{Macro}{Description of the macroscopic specimen, in text}
  \item{Diagnoses}{Description of the Diagnoses, in text}
}}
\usage{
Histop_df
}
\description{
A dataset containing fake pathology reports. The report fields have already been
separated into their own fields as opposed to PathDataFrameFinal and PathDataFrameFinalColon
}
\keyword{datasets}
