% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{HowManyTests}
\alias{HowManyTests}
\title{HowManyTests}
\usage{
HowManyTests(x, Indication, Endo_ResultPerformed, StringToSearch)
}
\arguments{
\item{x}{dataframe}

\item{Indication}{Indication column}

\item{Endo_ResultPerformed}{column containing date the Endoscopy was performed}

\item{StringToSearch}{The string in the Indication to search for}
}
\description{
Get an overall idea of how many endoscopies have been done for an indication by year and month
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v<-BarrettsDataAccord_Prague(v,"Endo_ResultText")
HowManyTests(v,"Indications","Endo_ResultPerformed","Surv")
}
\keyword{Tests}
\keyword{number}
