% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{LastStatus}
\alias{LastStatus}
\title{Last status}
\usage{
LastStatus(dataframe, HospNum, EVENT, indicatorEvent, endEvent)
}
\arguments{
\item{dataframe}{The dataframe}

\item{HospNum}{The Hospital Number column}

\item{EVENT}{The column  called EVENT that determines what procedure the
patient had at that endoscopy}

\item{indicatorEvent}{The name of the Event that references the outcome}

\item{endEvent}{The event that indicates the outcome has been reached}
}
\description{
This looks at the last EVENT a patient had to get time to outcomes.
For example if a patient underwent ablation for Barrett's oesophagus we
could get the CRIM score by assessing the first time-point where the event
is nothing after RFA has been done.
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-Mypath
v<-HistolDx(v,'Diagnosis')
v$NumBx<-HistolNumbOfBx(v$Macroscopicdescription,'specimen')
v$BxSize<-HistolBxSize(v$Macroscopicdescription)
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The function relies on the other Barrett's functions being run as well:
b1<-Barretts_PragueScore(v,'Findings')
b2<-Barretts_PathStage(b1,'Histology')
colnames(b2)[colnames(b2) == 'pHospitalNum'] <- 'HospitalNumber'
# The function groups the procedures by patient and then looks at those which
# have 'nothing' in the event column (which means biopsies only) which was
# preceded by radiofrequency ablation (RFA) so that these patients are
# labelled as having clearance of intestinal metaplasia. The result is a true
# or false column.
b2$EVENT<-EndoscopyEvent(b2,"Findings","ProcedurePerformed","Macroscopicdescription","Histology")
nn<-Barretts_CRIM(b2,'HospitalNumber',"EVENT")
rm(v)
}
\keyword{CRIM}
