% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{LocationListUniversal}
\alias{LocationListUniversal}
\title{Use list of standard locations for upper GI endoscopy}
\usage{
LocationListUniversal()
}
\description{
<<<<<<< HEAD
The is a list of standard locations at endoscopy that is used in the
extraction of the site of biopsies/EMRs and potentially in functions looking at the site of a
therapeutic event. It just returns the list in the function
=======
The is a list of standard locations at endoscopy that is used in the 
extraction of the site of biopsies/EMRs and potentially in functions looking at the site of a 
therapeutic event. It just returns the list in the function.
>>>>>>> master
}
\examples{
# No example needed
}
\keyword{Location}
