% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{LocationListUpper}
\alias{LocationListUpper}
\title{Use list of standard locations for upper GI endoscopy}
\usage{
LocationListUpper()
}
\description{
The is a list of standard locations at endoscopy that is used in the
extraction of the site of biopsies/EMRs and potentially in functions looking at the site of a 
therapeutic event. It just returns the list in the function.
}
\seealso{
Other NLP - Lexicons: \code{\link{BiopsyIndex}},
  \code{\link{EventList}}, \code{\link{GISymptomsList}},
  \code{\link{HistolType}},
  \code{\link{LocationListLower}},
  \code{\link{LocationListUniversal}},
  \code{\link{LocationList}}, \code{\link{RFACath}},
  \code{\link{WordsToNumbers}}
}
\concept{NLP - Lexicons}
\keyword{Location}
