% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{MetricByEndoscopist}
\alias{MetricByEndoscopist}
\title{Number of biopsies
This assess how many biopsies have been taken}
\usage{
MetricByEndoscopist(x, y, z)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column (numeric data) of interest}

\item{z}{The endoscopist column}
}
\description{
Number of biopsies
This assess how many biopsies have been taken
}
\examples{
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
Myendo<-EndoscChopperMeds(Myendo,'Medications')
Fent<-MetricByEndoscopist(Myendo,'Endoscopist','Fent')
}
\keyword{Endoscopist}
