% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Myendo}
\alias{Myendo}
\title{Fake Endoscopies}
\format{A data frame with 2000 rows and 1 variables:
\describe{
  \item{OGDReportWhole}{The whole report, in text}
  \item{HospitalNumber}{Hospital Number, in text}
  \item{PatientName}{Patient Name, in text}
  \item{GeneralPractitioner}{General Practitioner, in text}
  \item{Dateofprocedure}{Date of the procedure, as date}
  \item{Endoscopist}{Endoscopist, in text}
  \item{Secondendoscopist}{Secondendoscopist, in text}
  \item{Medications}{Medications, in text}
  \item{Instrument}{Instrument, in text}
  \item{ExtentofExam}{ExtentofExam, in text}
  \item{Indications}{Indications, in text}
  \item{ProcedurePerformed}{Procedure Performed, in text}
  \item{Findings}{Endoscopic findings, in text}
}}
\usage{
Myendo
}
\description{
A dataset containing fake endoscopy reports. The report fields have already been
The report field is derived from the whole report as follows:
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub('2nd Endoscopist:','Second endoscopist:',Myendo$OGDReportWhole)
EndoscTree<-list('Hospital Number:','Patient Name:','General Practitioner:',
'Date of procedure:','Endoscopist:','Second endoscopist:','Medications',
'Instrument','Extent of Exam:','Indications:','Procedure Performed:','Findings:',
'Endoscopic Diagnosis:')
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,'OGDReportWhole',as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
}
\keyword{datasets}
