% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{MyImgLibrary}
\alias{MyImgLibrary}
\title{Clean html endoscopic images}
\usage{
MyImgLibrary(file, delim, location)
}
\arguments{
\item{file}{The html report to extract (the html will have all the images references in it)}

\item{delim}{The phrase that separates individual endoscopies}

\item{location}{The folder containing the actual images}
}
\description{
This is used to pick and clean endoscopic images from html exports so they can be prepared
before being linked to pathology and endoscopy reports
}
\examples{
# MyImgLibrary("S:/Gastroenterology/Seb/R/Data/Barretts/Images Captured with Proc Data Audit_Findings1.html",
#                         "procedureperformed","S:/Gastroenterology/Seb/R/Data/Barretts/")
}
\seealso{
Other Basic Column mutators: 
\code{\link{EntityPairs_OneSentence}()},
\code{\link{EntityPairs_TwoSentence}()},
\code{\link{ExtrapolatefromDictionary}()},
\code{\link{ListLookup}()}
}
\concept{Basic Column mutators}
\keyword{Image}
\keyword{extraction}
