% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Mypath}
\alias{Mypath}
\title{Fake Pathology report}
\format{A data frame with 2000 rows and 1 variables:
\describe{
  \item{PathReportWhole}{The whole report, in text}
  \item{HospitalNumber}{Hospital Number, in text}
  \item{PatientName}{Patient Name, in text}
  \item{DOB}{Date of Birth, in text}
  \item{GeneralPractitioner}{General Practitioner, in text}
  \item{Dateofprocedure}{Date of the procedure, as date}
  \item{ClinicalDetails}{Clinical Details, in text}
  \item{Macroscopicdescription}{Macroscopic description of the report, in text}
  \item{Histology}{Histology, in text}
  \item{Diagnosis}{Diagnosis, in text}
}}
\usage{
Mypath
}
\description{
A dataset containing fake pathology reports.
The report field is derived from the whole report as follows:
Mypath<-PathDataFrameFinalColon
HistolTree<-list('Hospital Number','Patient Name','DOB:','General Practitioner:',
'Date of procedure:','Clinical Details:','Macroscopic description:','Histology:','Diagnosis:','')
for(i in 1:(length(HistolTree)-1)) {
 Mypath<-Extractor(Mypath,'PathReportWhole',as.character(HistolTree[i]),
 as.character(HistolTree[i+1]),as.character(HistolTree[i]))
}
Mypath$Dateofprocedure<-as.Date(Mypath$Dateofprocedure)
}
\keyword{datasets}
