% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{NegativeRemoveWrapper}
\alias{NegativeRemoveWrapper}
\title{Wrapper for Negative Remove}
\usage{
NegativeRemoveWrapper(inputText)
}
\value{
This returns a column within a dataframe. This should be changed to a 
character vector eventually
}
\description{
This performs negative removal on a per sentance basis
}
\examples{
# Build a character vector and then
# incorporate into a dataframe
anexample<-c("There is no evidence of polyp here",
"Although the prep was poor,there was no adenoma found",
"The colon was basically inflammed, but no polyp was seen",
"The Barrett's segment was not biopsied",
"The C0M7 stretch of Barrett's was flat")
anexample<-data.frame(anexample)
names(anexample)<-"Thecol"
# Run the function on the dataframe and it should get rid of sentences (and
# parts of sentences) with negative parts in them.
#hh<-NegativeRemoveWrapper(anexample$Thecol)
}
\seealso{
Other NLP - Text Cleaning and Extraction: 
\code{\link{ColumnCleanUp}()},
\code{\link{DictionaryInPlaceReplace}()},
\code{\link{Extractor}()},
\code{\link{NegativeRemove}()},
\code{\link{textPrep}()}
}
\concept{NLP - Text Cleaning and Extraction}
\keyword{Negative}
\keyword{Sentences}
