% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{NumberPerformed}
\alias{NumberPerformed}
\title{Determines the number of endoscopies done by an endoscopist by type of endosopy and indication for a given timeframe
As per BSG recommendations for Upper GI minimum number of gastroscopies in a year (although here the time frame is user
defined)}
\usage{
NumberPerformed(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column containing the Endoscopists names}

\item{z}{The column containing the Indication for the examination}
}
\description{
Determines the number of endoscopies done by an endoscopist by type of endosopy and indication for a given timeframe
As per BSG recommendations for Upper GI minimum number of gastroscopies in a year (although here the time frame is user
defined)
}
\examples{
v<-EndoRaw2()
 EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:","Date of procedure:","Endoscopist:","Endoscopist:","Medications","Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:","Endoscopic Diagnosis:")
 for(i in 1:(length(EndoscTree)-1)) {
 v<-Extractor(v,"OGDReportWhole",as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
 }
 v$Dateofprocedure<-as.Date(v$Dateofprocedure,format="\%d/\%m/\%Y")
 v<-EndoscChopperEndoscopist(v,"Endoscopist")
 v$Endoscopist<-gsub("Second","",v$Endoscopist)
 v<-EndoscChopperMeds(v,"Medications")
 v<-EndoscChopperInstrument(v,"Instrument")
 v<-EndoscChopperIndications(v,"Indications")
 v<-EndoscChopperProcPerformed(v,"ProcedurePerformed")
 v<-EndoscChopperFindings(v,"Findings")
 v<-v[grepl("Gastroscopy",v$ProcedurePerformed),]
}
\keyword{Withdrawal}
