% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PathDataFrameFinal}
\alias{PathDataFrameFinal}
\title{Fake Upper GI Pathology Set}
\format{A data frame with 2000 rows and 1 variables:
\describe{
  \item{PathReportWhole}{The whole report, in text}
}}
\usage{
PathDataFrameFinal
}
\description{
A dataset containing fake pathology reports for upper GI endoscopy tissue specimens.
The report field is provided as a whole report
without any fields having been already extracted
}
\keyword{datasets}
