% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{pathRep2}
\alias{pathRep2}
\title{pathRep}
\usage{
pathRep2()
}
\arguments{
\item{x}{None needed}
}
\description{
Creates raw Pathology reports
}
\examples{
pathRep(x)
}
\keyword{Pathology}
\keyword{reports}
