% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PatientFlow_CircosPlots}
\alias{PatientFlow_CircosPlots}
\title{Circos plot functions:}
\usage{
PatientFlow_CircosPlots(x, Endo_ResultPerformed, HospNum_Id, ProcPerformed)
}
\arguments{
\item{x}{dataframe}

\item{Endo_ResultPerformed}{the column containing the date of the procedure}

\item{ProcPerformed}{The procedure that you want to plot (eg EMR,
radiofrequency ablation for Barrett's but can be any dscription of a procedure you desire)}

\item{Indication}{the column containing the indication for the procedure}
}
\description{
This allows us to look at the overall flow from one type of procedure to another using circos plots
}
\keyword{Circos}
