% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PatientFlowBasic}
\alias{PatientFlowBasic}
\title{Determine the patient metric of choice over time WORK IN PROGRESS}
\usage{
PatientFlowBasic(theframe, EndoReportColumn, myNotableWords)
}
\arguments{
\item{theframe}{the dataframe,}

\item{EndoReportColumn}{the column of interest,}

\item{myNotableWords}{list of words you are interested in}
}
\description{
This function aims to show what is happening over time to a metric
of choice on a per patient basis.
An example might be to demonstrate the worst grade of histopathology
on repeated endoscopic biopsies eg for Barrett's oesophagus
}
\keyword{flow}
\keyword{patient}
