% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{PatientFlowIndividual}
\alias{PatientFlowIndividual}
\title{Create a plot over time of patient categorical findings as a line chart}
\usage{
PatientFlowIndividual(
  theframe,
  EndoReportColumn,
  myNotableWords,
  DateofProcedure,
  PatientID
)
}
\arguments{
\item{theframe}{dataframe}

\item{EndoReportColumn}{the column containing the date of the procedure}

\item{myNotableWords}{The terms from a column with categorical variables}

\item{DateofProcedure}{Column with the date of the procedure}

\item{PatientID}{Column with the patient's unique identifier}
}
\description{
This plots the findings at endoscopy (or pathology) over time for individual
patients. An example might be with worst pathological grade on biopsy for 
Barrett's oesophagus over time
}
\examples{
# This function builds chart of categorical outcomes for individal patients over time
# It allows a two dimensional visualisation of patient progress. A perfect example is 
# visualising the Barrett's progression for patients on surveillance and then
# therapy if dysplasia develops and highlighting recurrence if it happens
# Barretts_df <- BarrettsAll(Myendo, "Findings", "OGDReportWhole", Mypath, "Histology")
# myNotableWords<-c("No_IM","IM","LGD","HGD","T1a","IGD","SM1","SM2")
# PatientFlowIndividual(Barretts_df,"IMorNoIM",myNotableWords,DateofProcedure,"HospitalNumber")
# Once the function is run you should always call dev.off()
}
\seealso{
Other Patient Flow functions: 
\code{\link{SurveySankey}()}
}
\concept{Patient Flow functions}
\keyword{Patient}
\keyword{flow}
