% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{POS_Extract}
\alias{POS_Extract}
\title{Extract sentences with the POS tags desired}
\usage{
POS_Extract(POS_seq, columnPOS, columnSentence)
}
\arguments{
\item{POS_seq}{The POS tag sequence to extract}

\item{columnPOS}{The column with the POS tags}

\item{columnSentence}{The column with the sentence}

\item{dataframe}{The dataframe}
}
\description{
This uses udpipe to tag the text. It then compresses all of the text 
so you have continuous POS tagging or the whole text. The udpipe package has to be
pre downloaded to run this.
}
\examples{
#Has to be after POS extraction is done
#mylist<-POS_Extract("NOUN;ADJ;NOUN",MergedUp$upos,MergedUp$sentence)
#MergedUp$Extr<-mylist
}
\keyword{Macroscopic}
