% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{RFACath}
\alias{RFACath}
\title{Use list of catheters used in radiofrequency ablation}
\usage{
RFACath()
}
\description{
The takes a list of catheters used in radiofrequency ablation.
}
\seealso{
Other NLP - Lexicons: 
\code{\link{BiopsyIndex}()},
\code{\link{EventList}()},
\code{\link{GISymptomsList}()},
\code{\link{HistolType}()},
\code{\link{LocationListLower}()},
\code{\link{LocationListUniversal}()},
\code{\link{LocationListUpper}()},
\code{\link{LocationList}()},
\code{\link{WordsToNumbers}()}
}
\concept{NLP - Lexicons}
\keyword{RFA}
