% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SampleLocator}
\alias{SampleLocator}
\title{Locate where samples are taken from}
\usage{
SampleLocator(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{SampleLocation from the TermStandardizer}
}
\description{
This assess where samples are taken from.
This should be used after the TermStandardLocation
as it relies on the presence of a SampleLocation column
which the TermStandardLocation produces.
It can be used after the PolypTidyUpLocator
although you could just run it after the TermStandardLocation
to get an overview of where samples were taken from.
It will tell you the sites sampled without duplication
}
\examples{
Histoltree <-list("Hospital Number:","Patient Name:",
"General Practitioner:","Date received:","Clinical Details",
 "Nature of specimen","Histology","Diagnosis",""
)

for (i in 1:(length(Histoltree) - 1)) {
 Mypath <-Extractor(
     Mypath, "PathReportWhole",as.character(Histoltree[i]),
     as.character(Histoltree[i + 1]),
     gsub(" ", "", as.character(Histoltree[i]))
   )
}
names(Mypath)[names(Mypath) == 'Datereceived'] <- 'Dateofprocedure'
Mypath$Dateofprocedure <- as.Date(Mypath$Dateofprocedure)
f<-TermStandardLocation(Mypath,'Histology')
f<-SampleLocator(f,'SampleLocation')
}
\keyword{Sample}
\keyword{location}
