% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{#' Set the fill theme for all the ggplots
#'
#'}
\usage{
scale_fill_Publication(...)
}
\description{
#' Set the fill theme for all the ggplots
#'
#'
}
\examples{
# None needed
}
\keyword{ggplot}
\keyword{themes}
