% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{spellCheck}
\alias{spellCheck}
\title{Find and Replace function}
\usage{
spellCheck(pattern, replacement, x, fixed = FALSE, ...)
}
\value{
This returns a character vector
}
\description{
This is a helper function for finding and replacing from dictionaries like the event list
It uses fuzzy find and replace to account for spelling errors
}
\examples{
# Pending

}
\keyword{Find}
\keyword{and}
\keyword{replace}
