% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceCapacity}
\alias{SurveillanceCapacity}
\title{SurveillanceCapacity}
\usage{
SurveillanceCapacity(x, Endo_ResultPerformed)
}
\arguments{
\item{x}{dataframe}
}
\description{
This determines the last test done by that patient and the time between now and that last test
}
\examples{
SurveillanceFirstTest(v)
}
\keyword{cats}
