% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceFirstTest}
\alias{SurveillanceFirstTest}
\title{SurveillanceFirstTest}
\usage{
SurveillanceFirstTest(x, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{x}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
Extract the first test only per patient
}
\examples{
v<-TheOGDReportFinal
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v<-BarrettsDataAccord_Prague(v,'Endo_ResultText')
em<-SurveillanceFirstTest(v,'HospitalNumber','Endo_ResultPerformed')
}
\keyword{Surveillance}
