% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceTimeByRow}
\alias{SurveillanceTimeByRow}
\title{Extract the time difference between each test in days}
\usage{
SurveillanceTimeByRow(dataframe, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{dataframe,}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
This determines the time difference between each test for a patient in days.
}
\examples{
em<-SurveillanceTimeByRow(Myendo,'HospitalNumber',
'Dateofprocedure')
}
\keyword{Surveillance}
