% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveilTimeByRow}
\alias{SurveilTimeByRow}
\title{Extract the time difference between each test in days}
\usage{
SurveilTimeByRow(dataframe, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{dataframe,}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
This determines the time difference between each test for a patient in days
It returns the time since the first and the last study as a new dataframe.
}
\examples{
aa<-SurveilTimeByRow(Myendo,'HospitalNumber',
'Dateofprocedure')
}
\keyword{Surveillance}
