% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveySankey}
\alias{SurveySankey}
\title{SurveySankey}
\usage{
SurveySankey(dfw, y)
}
\arguments{
\item{y}{the column containing the test like ProcPerformed for example}

\item{x}{the dataframe extracted using the standard cleanup scripts}
}
\description{
This creates a Sankey plot to see the order of tests for all patients: 
dfw is the dataframe, 
y is the value of in this case the procedure type (eg EMR,
 radiofrequency ablation for Barrett's but can be any dscription of a procedure you desire)
 Note the Hospital Number column MUST be called PatientID
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v<-BarrettsDataAccord_Prague(v,"Endo_ResultText")
}
\keyword{Sankey}
