% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{Temporal}
\alias{Temporal}
\title{Temporal sentences}
\usage{
Temporal(inputString)
}
\arguments{
\item{inputString}{the string to be analysed}
}
\description{
This extracts sentences that have a tense to them so you can see
if there are intended actions for example in assessing quality
of a report eg 'I will prescribe this medication. The function
requires you to decide whether you want present, past of future tenses. You
can also choose to exclude tenses.
}
\examples{
pp<-udmodel_english <- udpipe_load_model(file = udmodel$file_model)
Temporal(Myendo$Findings)
}
\keyword{Macroscopic}
