% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{TemporalList}
\alias{TemporalList}
\title{Temporal list}
\usage{
TemporalList()
}
\description{
This function returns all the conversions from common version of events to
a standardised event list, much like the Location standardidastion function
It is used in the Barretts_EventType. This does not include EMR as this is
extracted from the pathology so is part of pathology type.
}
\examples{
# unique(unlist(Temporal(), use.names = FALSE))
}
\keyword{Temporal}
\keyword{extraction}
