% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{TemporalPOS}
\alias{TemporalPOS}
\title{Temporal sentences}
\usage{
TemporalPOS(x2)
}
\arguments{
\item{x2}{the dataframe with all the POS tags added}
}
\description{
This extracts sentences that have a tense to them so you can see
if there are intended actions for example in assessing quality
of a report eg 'I will prescribe this medication. The function
requires you to decide whether you want present, past of future tenses. You
can also choose to exclude tenses.
}
\examples{
# Run the EndoPOS example first 
Myendo2<-head(Myendo,100)
MyPOSframe<-EndoPOS(Myendo2$OGDReportWhole)
}
\keyword{Macroscopic}
