% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{TermStandardLocation}
\alias{TermStandardLocation}
\title{Standardises the location of biopsies
This assess the location of the biopies taken from the histopathology specimens.}
\usage{
TermStandardLocation(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{Column describing the Macroscopic sample from histology}
}
\description{
Standardises the location of biopsies
This assess the location of the biopies taken from the histopathology specimens.
}
\examples{
f<-TermStandardLocation(Mypath,'Histology')
}
\keyword{Withdrawal}
