% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{textPrep}
\alias{textPrep}
\title{textPrep function}
\usage{
textPrep(inputText)
}
\arguments{
\item{inputText}{The relevant pathology text column}
}
\value{
This returns a nested list
}
\description{
This is a helper function to prepare the data for the extraction of event, tissue type and site
from raw pathology and endoscopy files associated with the endoscopy. This is used within the 
BarrettsPathSite and BarrettsEventType function and is also used to generate OPCS-4 codes. Note
it needs a dataframe not a tibble so this needs to be converted prior to usage.
}
\examples{
CleanResults<-textPrep(TheOGDReportFinal$OGDReportWhole)

}
\keyword{Find}
\keyword{and}
\keyword{replace}
