% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{theme_Publication}
\alias{theme_Publication}
\title{#' Sets the publication theme for all the ggplots
#'
#'}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\description{
#' Sets the publication theme for all the ggplots
#'
#'
}
\examples{
#None needed
}
\keyword{ggplot}
\keyword{themes}
