% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{TimeToStatus}
\alias{TimeToStatus}
\title{Last status}
\usage{
TimeToStatus(dataframe, HospNum, EVENT, indicatorEvent, endEvent)
}
\arguments{
\item{dataframe}{The dataframe}

\item{HospNum}{The Hospital Number column}

\item{EVENT}{The column that contains the outcome of choice}

\item{indicatorEvent}{The name of the start event (can be a regular expression)}

\item{endEvent}{The name of the endpoint (can be a regular expression)}
}
\description{
This function selects patients who have had a start event and an end
event of the users choosing so you can determine things like how long
it takes to get a certain outcome. For example, how long does it take to
get a patient into a fully squamous oesophagus after Barrett's ablation
for dysplasia?
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v <- Mypath
v$NumBx <- HistolNumbOfBx(v$Macroscopicdescription, "specimen")
v$BxSize <- HistolBxSize(v$Macroscopicdescription)

# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v <- Endomerge2(
  Myendo, "Dateofprocedure", "HospitalNumber", v, "Dateofprocedure",
  "HospitalNumber"
)

# The function relies on the other Barrett's functions being run as well:
b1 <- Barretts_PragueScore(v, "Findings")
b1$IMorNoIM <- Barretts_PathStage(b1, "Histology")
colnames(b1)[colnames(b1) == "pHospitalNum"] <- "HospitalNumber"

# The function groups the procedures by patient and gives
# all the procedures between
<<<<<<< HEAD
# the indicatorEvent amd the procedure just after the endpoint.
# Eg if the start is RFA and the
# endpoint is biopsies then it will give all RFA procedures and
# the first biopsy procedure

b1$EndoscopyEvent <- EndoscopyEvent(
  b1, "Findings", "ProcedurePerformed",
  "Macroscopicdescription", "Histology"
)
nn <- TimeToStatus(b1, "HospitalNumber", "EndoscopyEvent", "rfa", "dilat")
=======
# the indicatorEvent amd the procedure just after the endpoint. 
# Eg if the start is RFA and the 
# endpoint is biopsies then it will give all RFA procedures and 
# the first biopsy procedure

b1$EndoscopyEvent<-EndoscopyEvent(b1,"Findings","ProcedurePerformed",
"Macroscopicdescription","Histology")
nn<-TimeToStatus(b1,'HospitalNumber',"EndoscopyEvent","rfa","dilat")
>>>>>>> master
rm(v)
}
\keyword{ourcome}
