% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{WordsToNumbers}
\alias{WordsToNumbers}
\title{Convetr words to numbers especially for the histopathology text}
\usage{
WordsToNumbers()
}
\description{
This function converts words to numbers.
}
\keyword{Event}
\keyword{extraction}
