% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_CRIM}
\alias{Barretts_CRIM}
\title{Barretts_CRIM}
\usage{
Barretts_CRIM(x, HospNum, EVENT)
}
\arguments{
\item{x}{The dataframe}

\item{HospNum}{The Hospital Number column}

\item{EVENT}{The column  called EVENT that determines what procedure the 
patient had at that endoscopy}
}
\description{
This collects the patients in whom it is assumed that complete clearance of intestinal
metasplasia has occurred (CRIM) after ablation therapy. This is done by collecting those
endoscopies where the last EVENT was equal to "nothing" when the patient had undergone 
radiofrequency ablation at some point
}
\keyword{CRIM}
