% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_LesionRecognitionEMR}
\alias{Barretts_LesionRecognitionEMR}
\title{Paris vs histopath Barrett's}
\usage{
Barretts_LesionRecognitionEMR(EndoSubsetEMR, y, z)
}
\arguments{
\item{EndoSubsetEMR}{The dataframe}

\item{y}{Endoscopy report field of interest}

\item{z}{Another endoscopy report field of interest}
}
\description{
This looks at the Paris grades of each EMR and then creates a heatmap 
of pathological grade vs
endoscopic Paris grade.This should only be run after all the 
BarrettsDataAccord functions.
}
\examples{
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Natureofspecimen','specimen')
v<-HistolChopperBxSize(v,'Natureofspecimen')
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
                             'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
Barretts_LesionRecognitionEMR(b4,"ProcedurePerformed","Findings")
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
