% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_LesionRecognitionEMR}
\alias{Barretts_LesionRecognitionEMR}
\title{Paris vs histopath Barrett's}
\usage{
Barretts_LesionRecognitionEMR(EndoSubsetEMR, y, z)
}
\arguments{
\item{EndoSubsetEMR}{The dataframe}

\item{y}{Endoscopy report field of interest}

\item{z}{Another endoscopy report field of interest}
}
\description{
This looks at the Paris grades of each EMR and then creates a heatmap 
of pathological grade vs
endoscopic Paris grade.This should only be run after all the 
BarrettsDataAccord functions.
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
