% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{Barretts_LesionRecognitionEMR}
\alias{Barretts_LesionRecognitionEMR}
\title{Paris vs histopath Barrett's}
\usage{
Barretts_LesionRecognitionEMR(EndoSubsetEMR, y, z)
}
\arguments{
\item{EndoSubsetEMR}{The dataframe}

\item{y}{Endoscopy report field of interest}

\item{z}{Another endoscopy report field of interest}
}
\description{
This looks at the Paris grades of each EMR and then creates a heatmap 
of pathological grade vs
endoscopic Paris grade.This should only be run after all the 
BarrettsDataAccord functions.
}
\examples{
# Firstly relevant columns are extrapolated from the 
# Mypath demo dataset. These functions are all part of Histology data 
# cleaning as part of the package. 
v<-HistolChopperAccessionNumber(Mypath,'Histology',
'SP-\\\\d{2}-\\\\d{7}')
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
#The function relies on the other Barrett's functions being run as well:
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
# The follow-up group depends on the histology and the Prague score for a 
# patient so it takes the processed Barrett's data and then looks in the 
# Findings column for permutations of the Prague score.
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function compares the Paris score from the endoscopy report free text to
# the histopathology scores for the same endoscopies so you can see what the
# lesion recognition is like overall
Barretts_LesionRecognitionEMR(b4,"ProcedurePerformed","Findings")
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
