% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsAll}
\alias{BarrettsAll}
\title{Follow up group determination}
\usage{
BarrettsAll(dataframe)
}
\arguments{
\item{dataframe}{the dataframe(merged endoscopy and histology).}
}
\description{
This function collects all the Barrett's functions togather as a 
parent function. The function checks the column names and performs
the necessary subfunction as long as the column is named correctly. 
The function follows the Histology function. You can use the same column
names as suggested with the parent function HistolAll
}
\examples{
 
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',Mypath,'Dateofprocedure',
'HospitalNumber')
#The function relies on the other Barrett's functions being run as well:
dd<-BarrettsAll(v)
rm(v)
}
\keyword{Barretts}
