% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsAll}
\alias{BarrettsAll}
\title{Follow up group determination}
\usage{
BarrettsAll(dataframe)
}
\arguments{
\item{dataframe}{the dataframe(merged endoscopy and histology).}
}
\description{
This function collects all the Barrett's functions togather as a 
parent function. The function checks the column names and performs
the necessary subfunction as long as the column is named correctly. 
The function follows the Histology function. You can use the same column
names as suggested with the parent function HistolAll
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
Mypath<-Extractor(Mypath,"PathReportWhole",mywords)
names(Mypath)<-c("Original","HospitalNumber","PatientName","DOB",
"GeneralPractitioner","Datereceived","ClinicalDetails",
"Macroscopicdescription","Histology","Diagnosis","HospitalNumber",
"PatientName","DOB","GeneralPractitioner","Dateofprocedure",
"ClinicalDetails","Macroscopicdescription","Histology","Diagnosis")
v<-HistolAll(Mypath)
rm(Mypath)

#' # Rename the columns in whatever endoscopy dataframe you have
names(Myendo)<-c("OGDReportWhole","HospitalNumber","PatientName",
"GeneralPractitioner","Dateofprocedure","Endoscopist","Secondendoscopist",
"Medications","Instrument","ExtentofExam","Indications","ProcedurePerformed",
"Findings" )
#Now use the function
Myendo<-EndoscAll(Myendo)

# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')


#The function relies on the other Barrett's functions being run as well:
BarrettsAll(v)
rm(v)
rm(Myendo)
}
\keyword{Barretts}
