% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Barretts.R
\name{BarrettsAll}
\alias{BarrettsAll}
\title{All basic Barrett's functions}
\usage{
BarrettsAll(Endodataframe, EndoReportColumn, EndoReportColumn2,
  Pathdataframe, PathColumn)
}
\arguments{
\item{Endodataframe}{endoscopy dataframe of interest}

\item{EndoReportColumn}{Endoscopy report field of interest as a string vector}

\item{EndoReportColumn2}{Second endoscopy report field of interest as a string vector}

\item{Pathdataframe}{pathology dataframe of interest}

\item{PathColumn}{Pathology report field of interest as a string vector}
}
\value{
Newdf
}
\description{
Function to encapsulate all the Barrett's functions together
}
\examples{
Barretts_df <- BarrettsAll(Myendo, "Findings", "OGDReportWhole", Mypath, "Histology")
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
