% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsBasicNumbers}
\alias{BarrettsBasicNumbers}
\title{Barretts Basic Numbers}
\usage{
BarrettsBasicNumbers(dataframe, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{the dataframe}

\item{Endo_ResultPerformed}{the date the endoscopy was performed}
}
\description{
This looks at the basic numbers of all the therapeutic endoscopies over time.
This should only be run after all the BarrettsDataAccord functions.
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolChopperDx(Mypath,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,
"Dateofprocedure","HospitalNumber")
# The function relies on the other Barrett's functions being run as well:
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
 'ProcedurePerformed','OGDReportWhole','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function groups the overall number of surveillance cases over time
# The endoscopic episodes should be selected according to surveillance
# being the indication prior to using this function
BarrettsBasicNumbers(b4,"Date.x")
rm(v)
}
\keyword{Number}
\keyword{of}
\keyword{therapies}
