% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsBasicNumbers}
\alias{BarrettsBasicNumbers}
\title{Barretts Basic Numbers}
\usage{
BarrettsBasicNumbers(x, Endo_ResultPerformed)
}
\arguments{
\item{x}{the dataframe}

\item{Endo_ResultPerformed}{the date the endoscopy was performed}
}
\description{
This looks at the basic numbers of all the therapeutic endoscopies over time.
This should only be run after all the BarrettsDataAccord functions.
}
\examples{
v<-HistolChopperDx(Mypath,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,
"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
                             'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsBasicNumbers(b4,"Date.x")
}
\keyword{Number}
\keyword{of}
\keyword{therapies}
