% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsBasicNumbers}
\alias{BarrettsBasicNumbers}
\title{Barretts Basic Numbers}
\usage{
BarrettsBasicNumbers(x, Endo_ResultPerformed)
}
\arguments{
\item{x}{the dataframe}

\item{Endo_ResultPerformed}{the date the endoscopy was performed}
}
\description{
This looks at the basic numbers of all the therapeutic endoscopies over time.
This should only be run after all the BarrettsDataAccord functions.
}
\keyword{Number}
\keyword{of}
\keyword{therapies}
