% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Event}
\alias{BarrettsDataAccord_Event}
\title{BarrettsDataAccord_Event}
\usage{
BarrettsDataAccord_Event(x, y, z, aa, bb)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The histology main text}

\item{z}{The Procedure Performed column}

\item{aa}{The endoscopic diagnosis column}

\item{bb}{The endoscopic findings column if different to the Diagnosis column}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it extracts the event
Being the procedure done at the time and the follow-up timings
}
\examples{
v<-TheOGDReportFinal
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'OGDReportWhole',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
b<-BarrettsDataAccord_Event(v,'Histo_ResultText','ProcedurePerformed','EndoscopicDiagnosis','Findings')

}
\keyword{Event}
\keyword{extraction}
