% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Event}
\alias{BarrettsDataAccord_Event}
\title{BarrettsDataAccord_Event}
\usage{
BarrettsDataAccord_Event(x, y, z, aa, bb)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The histology main text}

\item{z}{The Procedure Performed column}

\item{aa}{The endoscopic diagnosis column}

\item{bb}{The endoscopic findings column if different to the Diagnosis column}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it extracts the event
Being the procedure done at the time and the follow-up timings
}
\examples{
v<-HistolChopperAccessionNumber(Mypath,"Histology","SP-\\\\d{2}-\\\\d{7}")
v<-HistolChopperDx(v,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b<-BarrettsDataAccord_Event(v,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')

}
\keyword{Event}
\keyword{extraction}
