% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Event}
\alias{BarrettsDataAccord_Event}
\title{Therapeutic subtypes}
\usage{
BarrettsDataAccord_Event(x, y, z, aa, bb)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The histology main text}

\item{z}{The Procedure Performed column}

\item{aa}{The endoscopic diagnosis column}

\item{bb}{The endoscopic findings column if different to the Diagnosis column}
}
\description{
This function extracts the Event- usually a therapeutic event, from the text 
eg endoscopic mucosal resection, radiofrequency ablation etc. 
It does not currently include stricture
dilatation.Specfically it extracts the event
}
\examples{
# Firstly relevant columns are extrapolated from the 
# Mypath demo dataset. These functions are all part of Histology data 
# cleaning as part of the package. 

v<-HistolChopperAccessionNumber(Mypath,'Histology',
'SP-\\\\d{2}-\\\\d{7}')
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The function then looks within the Histology and the 
# Procedure performed, free text endoscopic Findings and the original 
# whole endoscopy report columns from the merged data set (v) for 
# EMR/APC/RFA/HALO so that the event for the procedure is recorded.
b<-BarrettsDataAccord_Event(v,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')

}
\keyword{Event}
\keyword{extraction}
