% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Event}
\alias{BarrettsDataAccord_Event}
\title{BarrettsDataAccord_Event}
\usage{
BarrettsDataAccord_Event(x, y, z, aa, bb)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The histology main text}

\item{z}{The Procedure Performed column}

\item{aa}{The endoscopic diagnosis column}

\item{bb}{The endoscopic findings column if different to the Diagnosis column}
}
\description{
This function extracts the Event- usually a therapeutic event, from the text 
eg endoscopic mucosal resection, radiofrequency ablation etc. 
It does not currently include stricture
dilatation.Specfically it extracts the event
}
\examples{
v<-HistolChopperAccessionNumber(Mypath,'Histology',
'SP-\\\\d{2}-\\\\d{7}')
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
b<-BarrettsDataAccord_Event(v,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')

}
\keyword{Event}
\keyword{extraction}
