% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Event}
\alias{BarrettsDataAccord_Event}
\title{BarrettsDataAccord_Event}
\usage{
BarrettsDataAccord_Event(x, y, z, aa, bb)
}
\arguments{
\item{x}{the dataframe}

\item{y}{The histology main text}

\item{z}{The Procedure Performed column}

\item{aa}{The endoscopic diagnpsis column}

\item{bb}{The endoscopic findings column if different to the Diagnosis column}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it extracts the event
Being the procedure done at the time and the follow-up timings
}
\examples{
BarrettsDataAccord_Event(b,'Histo_ResultText','PROCEDURE','DIAGNOSIS','FINDINGS')
}
\keyword{Event}
\keyword{extraction}
