% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_FUGroup}
\alias{BarrettsDataAccord_FUGroup}
\title{BarrettsDataAccord_FUGroup}
\usage{
BarrettsDataAccord_FUGroup(x, y)
}
\arguments{
\item{x}{the dataframe(which has to have been processed by the BarrettsDataAccord_PathStage function first to get IMorNoIM),}

\item{Findings}{The field to search (endoscopic findings)}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it combines the presence of intestinal metaplasia with Prague score so the follow-up group can be determined
Being the procedure done at the time and the follow-up timings
}
\examples{
BarrettsDataAccord_FUGroup(b,'FINDINGS')
}
\keyword{Follow-Up}
