% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_FUGroup}
\alias{BarrettsDataAccord_FUGroup}
\title{BarrettsDataAccord_FUGroup}
\usage{
BarrettsDataAccord_FUGroup(x, y)
}
\arguments{
\item{x}{the dataframe(which has to have been processed by the BarrettsDataAccord_PathStage function first to get IMorNoIM),}

\item{y}{The field to search (endoscopic findings)}
}
\description{
This function accordionises the data so we get all the bits extrapolated from what's available
Specfically it combines the presence of intestinal metaplasia with Prague score so the follow-up group can be determined
Being the procedure done at the time and the follow-up timings
}
\examples{
v<-TheOGDReportFinal
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
b<-BarrettsDataAccord_FUGroup(b,'Findings')
}
\keyword{Follow-Up}
