% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_PathStage}
\alias{BarrettsDataAccord_PathStage}
\title{BarrettsDataAccord_PathStage}
\usage{
BarrettsDataAccord_PathStage(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This extracts the pathological stage from the histopathology specimen. It is 
done using 'degradation' so that it will look for the worst overall grade 
in the histology specimen and if not found it will look for the next worst
and so on. It looks per report not per biopsy (it is more common
for histopathology reports tocontain the worst overall grade 
rather than individual biopsy grades).
Specfically it extracts the histopathology worst grade
Being the procedure done at the time and the follow-up timings
}
\examples{
v<-HistolChopperAccessionNumber(Mypath,'Histology','SP-\\\\d{2}-\\\\d{7}')
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
b<-BarrettsDataAccord_PathStage(v,'Histology')
}
\keyword{Pathology}
\keyword{extraction}
