% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsDataAccord_Prague}
\alias{BarrettsDataAccord_Prague}
\title{Prague score extraction}
\usage{
BarrettsDataAccord_Prague(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
The aim is to extract a C and M stage (Prague score) for Barrett's samples.
This is done using a regex where C and M stages are explicitly mentioned in
the free text
Specfically it extracts the Prague score
}
\examples{
v<-BarrettsDataAccord_Prague(Myendo,'Findings')
}
\keyword{Prague}
\keyword{score}
