% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Barretts.R
\name{BarrettsParisEMR}
\alias{BarrettsParisEMR}
\title{Run the Paris classification versus worst histopath grade for Barrett's}
\usage{
BarrettsParisEMR(Column, Column2)
}
\arguments{
\item{Column}{Endoscopy report field of interest as a string vector}

\item{Column2}{Another endoscopy report field of interest as a string vector}
}
\value{
a string vector
}
\description{
This creates a column of Paris grade for all samples where this is mentioned.
}
\examples{
# 
Myendo$EMR<-BarrettsParisEMR(Myendo$ProcedurePerformed,Myendo$Findings)
}
\seealso{
Other Disease Specific Analysis - Barretts Data: \code{\link{BarrettsAll}},
  \code{\link{BarrettsBxQual}},
  \code{\link{Barretts_FUType}},
  \code{\link{Barretts_PathStage}},
  \code{\link{Barretts_PragueScore}}
}
\concept{Disease Specific Analysis - Barretts Data}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
