% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Barretts.R
\name{BarrettsParisEMR}
\alias{BarrettsParisEMR}
\title{Paris vs histopath Barrett's}
\usage{
BarrettsParisEMR(Column, Column2)
}
\arguments{
\item{Column}{Endoscopy report field of interest as a string vector}

\item{Column2}{Another endoscopy report field of interest as a string vector}
}
\value{
a string vector
}
\description{
This creates a column of Paris grade for all samples where this is mentioned.
}
\examples{
<<<<<<< HEAD
#
Myendo$EMR <- BarrettsParisEMR(Myendo$ProcedurePerformed, Myendo$Findings)
rm(v)
=======
# 
Myendo$EMR<-BarrettsParisEMR(Myendo$ProcedurePerformed,Myendo$Findings)
>>>>>>> master
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
