% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPatientTracking_Enrollment_Surveillance}
\alias{BarrettsPatientTracking_Enrollment_Surveillance}
\title{BarrettsPatientTracking_Enrollment_Surveillance}
\usage{
BarrettsPatientTracking_Enrollment_Surveillance(x, PatientID,
  Endo_ResultPerformed, IndicationsFroExamination)
}
\arguments{
\item{x}{dataframe}

\item{PatientID}{column of interest with unique hospital number in it}

\item{Endo_ResultPerformed}{column of interest with date endiscopy performed in it}

\item{IndicationsFroExamination}{column of interest with indications in it (usually 'Surveillance' or similar)}
}
\description{
This function graphs the patients who were not on surveillance programmes and sees how many then had an endoscopy.
}
\examples{
BarrettsPatientTracking_Enrollment_Surveillance(b,'PatientID','Endo_ResultPerformed','INDIC')
}
\keyword{Patient}
\keyword{Tracking}
