% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPatientTracking_Enrollment_Surveillance}
\alias{BarrettsPatientTracking_Enrollment_Surveillance}
\title{BarrettsPatientTracking_Enrollment_Surveillance}
\usage{
BarrettsPatientTracking_Enrollment_Surveillance(x, PatientID,
  Endo_ResultPerformed, IndicationsFroExamination)
}
\arguments{
\item{x}{dataframe}

\item{PatientID}{column of interest with unique hospital number in it}

\item{Endo_ResultPerformed}{column of interest with date endiscopy performed in it}

\item{IndicationsFroExamination}{column of interest with indications in it (usually 'Surveillance' or similar)}
}
\description{
This function graphs the patients who were not on surveillance programmes and sees how many then had an endoscopy.
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
Enroll<-BarrettsPatientTracking_Enrollment_Surveillance(b,"HospitalNumber","Endo_ResultPerformed","Indications")
}
\keyword{Patient}
\keyword{Tracking}
