% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPatientTracking_UniqueHospNum}
\alias{BarrettsPatientTracking_UniqueHospNum}
\title{BarrettsPatientTracking_UniqueHospNum}
\usage{
BarrettsPatientTracking_UniqueHospNum(x, rule, PatientID)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{rule}{Rule of interest}

\item{PatientID}{Column containing patient numbers}
}
\description{
This function gets the unique patient ID's for each patient, for each rule. It lists the unique PatientIDs assocaited with a rule ('Rule1','Rule2','Rule3','NoRules')
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
Rule<-BarrettsPatientTracking_UniqueHospNum(b,"Rule1","HospNumId")
}
\keyword{Rule}
