% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPatientTracking_UniqueHospNum}
\alias{BarrettsPatientTracking_UniqueHospNum}
\title{BarrettsPatientTracking_UniqueHospNum}
\usage{
BarrettsPatientTracking_UniqueHospNum(x, rule, PatientID)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{rule}{Rule of interest}

\item{PatientID}{Column containing patient numbers}
}
\description{
This function gets the unique patient ID's for each patient, for each rule. It lists the unique PatientIDs assocaited with a rule ('Rule1','Rule2','Rule3','NoRules')
}
\examples{
v<-TheOGDReportFinal
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second Endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
Mypath<-PathDataFrameFinalColon
HistolTree<-list("Hospital Number","Patient Name","DOB:","General Practitioner:",
"Date of procedure:","Clinical Details:","Macroscopic description:","Histology:","Diagnosis:","")
for(i in 1:(length(HistolTree)-1)) {
 Mypath<-Extractor(Mypath,"PathReportWhole",as.character(HistolTree[i]),
 as.character(HistolTree[i+1]),as.character(HistolTree[i]))
}
Mypath$Dateofprocedure<-as.Date(Mypath$Dateofprocedure)
v<-HistolChopperAccessionNumber(Mypath,"Histology","SP-\\\\d{2}-\\\\d{7}")
v<-HistolChopperDx(v,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
Rule<-BarrettsPatientTracking_UniqueHospNum(b4,'Rule1','HospitalNumber')
}
\keyword{Rule}
