% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsPatientTracking_UniqueHospNum}
\alias{BarrettsPatientTracking_UniqueHospNum}
\title{BarrettsPatientTracking_UniqueHospNum}
\usage{
BarrettsPatientTracking_UniqueHospNum(x, rule, PatientID)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{rule}{Rule of interest}

\item{PatientID}{Column containing patient numbers}
}
\description{
This function gets the unique patient ID's for each patient, 
for each rule. It lists the unique PatientIDs assocaited with a rule 
('Rule1','Rule2','Rule3','NoRules'). This allows us to determine how many patients will need 
follow up at specific time intervals. This should be run after the BarrettsDataAccord_Prague 
and BarrettsDataAccord_PathStage.
}
\examples{
v<-HistolChopperAccessionNumber(Mypath,'Histology','SP-\\\\d{2}-\\\\d{7}')
v<-HistolChopperDx(v,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure','HospitalNumber')
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
Rule<-BarrettsPatientTracking_UniqueHospNum(b4,'Rule1','HospitalNumber')
}
\keyword{Rule}
