% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsQuality_AnalysisBiopsyNumber}
\alias{BarrettsQuality_AnalysisBiopsyNumber}
\title{BarrettsQuality_AnalysisBiopsyNumber}
\usage{
BarrettsQuality_AnalysisBiopsyNumber(x, Endo_ResultPerformed, PatientID,
  Endoscopist)
}
\arguments{
\item{x}{dataframe}

\item{Endo_ResultPerformed}{Date of the Endocscopy}

\item{PatientID}{Patient's unique identifier}

\item{Endoscopist}{name of the column with the Endoscopist names}
}
\description{
This function gets the biopsies taken per endoscopy and compares to the
Barrett's length
}
\examples{
BarrettsQuality_AnalysisBiopsyNumber(b,'Endo_ResultPerformed','PatientID','Endoscopist')
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
