% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsQuality_AnalysisDocumentation}
\alias{BarrettsQuality_AnalysisDocumentation}
\title{BarrettsQuality_AnalysisDocumentation}
\usage{
BarrettsQuality_AnalysisDocumentation(x, Findings)
}
\arguments{
\item{x}{dataframe}

\item{y}{column of interest- usually the main body of the endoscopic report}
}
\description{
This function assesses the Barrett's documentation. Need to have run the BarrettsDataAccord_Prague first
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
BarrettsQuality_AnalysisDocumentation(b,"Findings")
}
\keyword{Documentation}
