% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsQuality_AnalysisDocumentation}
\alias{BarrettsQuality_AnalysisDocumentation}
\title{BarrettsQuality_AnalysisDocumentation}
\usage{
BarrettsQuality_AnalysisDocumentation(x, Findings)
}
\arguments{
\item{x}{dataframe}

\item{Findings}{column of interest- usually the main body of the endoscopic report}
}
\description{
This function assesses the Barrett's documentation. Need to have run the BarrettsDataAccord_Prague first
}
\examples{
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b<-BarrettsDataAccord_PathStage(v,'Histology')
BarrettsQuality_AnalysisDocumentation(b,"Findings")
}
\keyword{Documentation}
