% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillance_PathDetection}
\alias{BarrettsSurveillance_PathDetection}
\title{BarrettsSurveillance_PathDetection}
\usage{
BarrettsSurveillance_PathDetection(x, titlePlot)
}
\arguments{
\item{x}{the dataframe}

\item{titlePlot}{The plot title}
}
\description{
This function assesses pathology of specimens taken at surveillance per year
}
\examples{
v<-Endomerge2(PathDataFrameFinal,TheOGDReportFinal)
names(v)<-c('Endo_ResultText','Endo_ResultPerformed','HospNumId',
'Histo_ResultText','Histo_ResultPerformed','HospNumId2','Days')
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
b<-BarrettsDataAccord_PathStage(v,'Histo_ResultText')
BarrettsSurveillance_PathDetection(b,'Myplot')
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
