% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillance_PathDetection}
\alias{BarrettsSurveillance_PathDetection}
\title{BarrettsSurveillance_PathDetection}
\usage{
BarrettsSurveillance_PathDetection(x, titlePlot)
}
\arguments{
\item{x}{the dataframe}

\item{titlePlot}{The plot title}
}
\description{
This function assesses pathology of specimens taken at surveillance per year
}
\examples{
v<-TheOGDReportFinal
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second Endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
Mypath<-PathDataFrameFinalColon
HistolTree<-list("Hospital Number","Patient Name","DOB:","General Practitioner:",
"Date of procedure:","Clinical Details:","Macroscopic description:","Histology:","Diagnosis:","")
for(i in 1:(length(HistolTree)-1)) {
 Mypath<-Extractor(Mypath,"PathReportWhole",as.character(HistolTree[i]),
 as.character(HistolTree[i+1]),as.character(HistolTree[i]))
}
Mypath$Dateofprocedure<-as.Date(Mypath$Dateofprocedure)
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsSurveillance_PathDetection(b4,'Myplot')
}
\keyword{Does}
\keyword{data}
\keyword{something}
\keyword{with}
