% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillance_PathDetection}
\alias{BarrettsSurveillance_PathDetection}
\title{Dysplasia detection on surveillance}
\usage{
BarrettsSurveillance_PathDetection(x, titlePlot)
}
\arguments{
\item{x}{the dataframe}

\item{titlePlot}{The plot title}
}
\description{
This function assesses pathology of specimens taken at surveillance per year.
It outputs a plot which determines the the overall number of pathologies (
low/high grade dysplasia and cancer) for patients on surveillance
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolChopperDx(Mypath,'Diagnosis')
v<-HistolChopperExtrapolDx(v,'Diagnosis')
v<-HistolChopperNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolChopperBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date and Hospital number
v<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',v,'Dateofprocedure',
'HospitalNumber')
# The function relies on the other Barrett's functions being run as well:
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function simply the the histopathological grades overall for
# your dataset and then creates a frequency plot of them
BarrettsSurveillance_PathDetection(b4,'Proportion with dysplasia or cancer')
rm(v)
}
\keyword{and}
\keyword{cancer}
\keyword{detection}
\keyword{dysplasia}
