% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillance_PathDetection}
\alias{BarrettsSurveillance_PathDetection}
\title{BarrettsSurveillance_PathDetection}
\usage{
BarrettsSurveillance_PathDetection(x, titlePlot)
}
\arguments{
\item{x}{the dataframe}

\item{titlePlot}{The plot title}
}
\description{
This function assesses pathology of specimens taken at surveillance per year.
It outputs a plot which determines the the overall number of pathologies (
low/high grade dysplasia and cancer) for patients on surveillance
}
\examples{
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsSurveillance_PathDetection(b4,'Myplot')
}
\keyword{and}
\keyword{cancer}
\keyword{detection}
\keyword{dysplasia}
