% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillanceDDR}
\alias{BarrettsSurveillanceDDR}
\title{BarrettsSurveillanceDDR}
\usage{
BarrettsSurveillanceDDR(x, y, IMorNoIM)
}
\arguments{
\item{x}{dataframe,}

\item{y}{column with the Endoscopist names}

\item{IMorNoIM}{extracted with the function BarrettsDataAccord_PathStage()}
}
\description{
This function assesses the dysplasia detection rate per endoscopist
}
\examples{
v<-TheOGDReportFinal
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second Endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
Mypath<-PathDataFrameFinalColon
HistolTree<-list("Hospital Number","Patient Name","DOB:","General Practitioner:",
"Date of procedure:","Clinical Details:","Macroscopic description:","Histology:","Diagnosis:","")
for(i in 1:(length(HistolTree)-1)) {
 Mypath<-Extractor(Mypath,"PathReportWhole",as.character(HistolTree[i]),
 as.character(HistolTree[i+1]),as.character(HistolTree[i]))
}
Mypath$Dateofprocedure<-as.Date(Mypath$Dateofprocedure)
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsSurveillanceDDR(b4,'Endoscopist','IMorNoIM')
}
\keyword{detection}
\keyword{dysplasia}
\keyword{rate}
