% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillanceDDR}
\alias{BarrettsSurveillanceDDR}
\title{BarrettsSurveillanceDDR}
\usage{
BarrettsSurveillanceDDR(x, y, z)
}
\arguments{
\item{x}{dataframe,}

\item{y}{column with the Endoscopist names}

\item{z}{column of the Pathological grade of specimen- is the column}

\item{IMorNoIM}{extracted with the function BarrettsDataAccord_PathStage()}
}
\description{
This function assesses the dysplasia detection rate per endoscopist
}
\examples{
BarrettsSurveillanceDDR(x,'Endoscopist','IMorNoIM')
}
\keyword{detection}
\keyword{dysplasia}
\keyword{rate}
