% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsSurveillanceDDR}
\alias{BarrettsSurveillanceDDR}
\title{BarrettsSurveillanceDDR}
\usage{
BarrettsSurveillanceDDR(x, y, IMorNoIM)
}
\arguments{
\item{x}{dataframe,}

\item{y}{column with the Endoscopist names}

\item{IMorNoIM}{extracted with the function BarrettsDataAccord_PathStage()}
}
\description{
This function assesses the dysplasia detection rate per endoscopist
}
\examples{
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
BarrettsSurveillanceDDR(b4,'Endoscopist','IMorNoIM')
}
\keyword{detection}
\keyword{dysplasia}
\keyword{rate}
