% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapeuticsOnly}
\alias{BarrettsTherapeuticsOnly}
\title{BarrettsTherapeuticsOnly}
\usage{
BarrettsTherapeuticsOnly(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function makes all therapeutics unique. 
This should only be run after the
BarrettsDataAccord_EventFunction as this provides an EVENT column which this
function uses.
}
\keyword{Therapeutic}
