% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapeuticsRFA_ByCatheter}
\alias{BarrettsTherapeuticsRFA_ByCatheter}
\title{RFA catheter use}
\usage{
BarrettsTherapeuticsRFA_ByCatheter(EndoSubsetRFA, Column, Column2)
}
\arguments{
\item{EndoSubsetRFA}{The dataframe}

\item{Column}{report field of interest}

\item{Column2}{Another endoscopy report field of interest}
}
\description{
This looks at the basic numbers of RFA by catheter type used.
This should only be run after all the BarrettsDataAccord functions.
}
\examples{
# Firstly relevant columns are extrapolated from the
# Mypath demo dataset. These functions are all part of Histology data
# cleaning as part of the package.
v<-HistolDx(Mypath,'Diagnosis')
v<-HistolExtrapolDx(v,'Diagnosis')
v<-HistolNumbOfBx(v,'Macroscopicdescription','specimen')
v<-HistolBxSize(v,'Macroscopicdescription')
# The histology is then merged with the Endoscopy dataset. The merge occurs
# according to date #and Hospital number
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,
"Dateofprocedure","HospitalNumber")
# The function relies on the other Barrett's functions being run as well:
b1<-BarrettsDataAccord_Prague(v,'Findings')
b2<-BarrettsDataAccord_PathStage(b1,'Histology')
b3<-BarrettsDataAccord_Event(b2,'Histology',
'ProcedurePerformed','OGDReportWhole','Findings')
# The follow-up group depends on the histology and the Prague score for a
# patient so it takes the processed Barrett's data and then looks in the
# Findings column for permutations of the Prague score.
b4<-BarrettsDataAccord_FUGroup(b3,'Findings')
colnames(b4)[colnames(b4) == 'pHospitalNum'] <- 'HospitalNumber'
# The function takes the RFA cases by looking in any free text column where
# endoscopy fingings are described and then summarising by RFA subtype
BarrettsTherapeuticsRFA_ByCatheter(b4,"ProcedurePerformed","Findings")
rm(v)
}
\keyword{RFA,}
\keyword{Radiofrequency}
\keyword{ablation}
