% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapeuticsRFA_ByCatheter}
\alias{BarrettsTherapeuticsRFA_ByCatheter}
\title{BarrettsTherapeuticsRFA_ByCatheter}
\usage{
BarrettsTherapeuticsRFA_ByCatheter(EndoSubsetRFA, y, z)
}
\arguments{
\item{EndoSubsetRFA}{The dataframe}

\item{y}{report field of interest}

\item{z}{Another endoscopy report field of interest}
}
\description{
This looks at the basic numbers of RFA by catheter type used
}
\keyword{RFA,}
\keyword{Radiofrequency}
\keyword{ablation}
