% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarrettsFunctions.R
\name{BarrettsTherapy_Numbers_EMRsByGrade}
\alias{BarrettsTherapy_Numbers_EMRsByGrade}
\title{BarrettsTherapy_Numbers_EMRsByGrade}
\usage{
BarrettsTherapy_Numbers_EMRsByGrade(EndoSubsetEMR)
}
\arguments{
\item{EndoSubsetEMR}{The dataset. Need to have run BarettsDataAccord_Event() 
and BarettsDataAccord_PathStage() first}
}
\description{
Plots all the pathological grades of the EMRs
}
\examples{
BarrettsTherapy_Numbers_EMRsByGrade(EndoSubsetEMR)
}
\keyword{EMR}
\keyword{chart}
