% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ChopperNewLines}
\alias{ChopperNewLines}
\title{Cleans endoscopy reports and splits sentences into new lines}
\usage{
ChopperNewLines(dataframe, EndoReportColumn)
}
\arguments{
\item{dataframe}{dataframe}

\item{EndoReportColumn}{The endoscopy report text column}
}
\description{
Cleans the long lines to separate them into new lines if necessary.
This helps further down the line in the tokenisation of the text.
It is optional depending on the format of your text. It should be used
after the Extractor has separated out the different
parts of the text according to the user's requirements
}
\examples{
# The function takes the demo data
# Myendo and searches through the raw
# endoscopy text so that the text is
#divided by sentence into a newline
v<-ChopperNewLines(Myendo,'OGDReportWhole')
}
\keyword{Endoscopy}
\keyword{Newlines}
