% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ChopperNewLines}
\alias{ChopperNewLines}
\title{EndoscChopperNewLines}
\usage{
ChopperNewLines(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{The endoscopy report text column}
}
\description{
Cleans the long lines to separate them into new lines if necessary
}
\examples{
v<-TheOGDReportFinal
EndoscTree<-list('Hospital Number:','Patient Name:','General Practitioner:',
'Date of procedure:','Endoscopist:','Endoscopist:','Medications','Instrument',
'Extent of Exam:','Indications:','Procedure Performed:','Findings:','Endoscopic Diagnosis:')
v<-ChopperNewLines(v,'OGDReportWhole')
}
\keyword{Endoscopy}
\keyword{Newlines}
