% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ColumnCleanUp}
\alias{ColumnCleanUp}
\title{Tidies up messy columns}
\usage{
ColumnCleanUp(dataframe, Column)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{Column}{column of interest}
}
\description{
This does a general clean up of whitespace,
semi-colons,full stops at the start
of lines and converts end sentence full stops to new lines.
It should be used after the Extractor.
It is used for columns where there is a lot of free text to extract. It
really extracts and standardises the sentences.
}
\examples{
ii<-ColumnCleanUp(Myendo,"Findings")
}
\keyword{Cleaner}
