% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{ColumnCleanUp}
\alias{ColumnCleanUp}
\title{ColumnCleanUp}
\usage{
ColumnCleanUp(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This does a general clean up of whitespace,semi-colons,full stops at the start
of lines and converts end sentence full stops to new lines
}
\examples{
 x[, y] <- ColumnCleanUp(x, y)
}
\keyword{Cleaner}
