% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{DictionaryInPlaceReplace}
\alias{DictionaryInPlaceReplace}
\title{Dictionary In Place Replace}
\usage{
DictionaryInPlaceReplace(inputString, list)
}
\arguments{
\item{inputString}{the input string}

\item{list}{The replacing list}
}
\value{
This returns a character vector
}
\description{
Standardises terms according to a user defined list and replaces them in place
}
\examples{
#Needs an example
}
\keyword{Replace}
