% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{DictionaryInPlaceReplace}
\alias{DictionaryInPlaceReplace}
\title{Dictionary In Place Replace}
\usage{
DictionaryInPlaceReplace(inputString, list)
}
\arguments{
\item{inputString}{the input string (ie the full medical report)}

\item{list}{The replacing list}
}
\value{
This returns a character vector
}
\description{
This maps terms in the text and replaces them with the 
standardised term (mapped in the lexicon file) within the text.
It is used within the textPrep function.
}
\examples{
inputText<-DictionaryInPlaceReplace(TheOGDReportFinal$OGDReportWhole,LocationList())
}
\keyword{Replace}
