% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoBasicGraph}
\alias{EndoBasicGraph}
\title{Basic graph creation using the template specified in theme_Publication.}
\usage{
EndoBasicGraph(dataframe, xdata, number)
}
\arguments{
\item{dataframe}{dataframe}

\item{xdata}{The Time column}

\item{number}{The numeric column}
}
\value{
Myplot the EDA final plot

Myplot
}
\description{
This creates a basic graph using the template specified in theme_Publication.
}
\examples{
# This function plot numeric y vs non-numeric x
# Get some numeric columns eg number of biopsies and size
Mypath$Size <- HistolBxSize(Mypath$Macroscopicdescription)
Mypath$NumBx <- HistolNumbOfBx(Mypath$Macroscopicdescription, "specimen")
Mypath2 <- Mypath[, c("NumBx", "Size")]
EndoBasicGraph(Mypath, "Size", "NumBx")
}
\keyword{Time}
\keyword{plots}
