% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{EndoDataVizNum_xNum_y}
\alias{EndoDataVizNum_xNum_y}
\title{Create  Numeric x versus y}
\usage{
EndoDataVizNum_xNum_y(dataframe, Numx, Numy)
}
\arguments{
\item{dataframe}{dataframe}

\item{Numx}{The numeric x column}

\item{Numy}{The numeric y column}
}
\value{
Myplot
}
\description{
Create  Numeric x versus y
}
\examples{
# This function plot numeric y vs non-numeric x


}
