% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{Endomerge2}
\alias{Endomerge2}
\title{Endomerge}
\usage{
Endomerge2(x, y)
}
\arguments{
\item{x}{Endoscopy dataframe}

\item{y}{Histopathology dataframe}
}
\description{
Merges endoscopy and histology data
}
\examples{
Endomerge(EndoscopyData,HistopathologyData)
}
\keyword{and}
\keyword{endoscopy}
\keyword{histology}
\keyword{merge}
