% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{Endomerge2}
\alias{Endomerge2}
\title{Merge endoscopy and histology data.}
\usage{
Endomerge2(x, EndoDate, EndoHospNumber, y, PathDate, PathHospNumber)
}
\arguments{
\item{x}{Endoscopy dataframe}

\item{EndoDate}{The date the endoscopy was performed}

\item{EndoHospNumber}{The unique hospital number in the endoscopy dataset}

\item{y}{Histopathology dataframe}

\item{PathDate}{The date the endoscopy was performed}

\item{PathHospNumber}{The unique hospital number in the endoscopy dataset}
}
\description{
This takes the endoscopy dataset date
performed and the hospital number column
and merges with the equivalent column in the pathology dataset. This is
merged within a 7 day time frame as pathology is often reported after
endoscopic
}
\examples{
v <- Endomerge2(
  Myendo, "Dateofprocedure", "HospitalNumber",
  Mypath, "Dateofprocedure", "HospitalNumber"
)
}
\keyword{and}
\keyword{endoscopy}
\keyword{histology}
\keyword{merge}
