% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoPOS}
\alias{EndoPOS}
\title{Parts of speech tagging for reports}
\usage{
EndoPOS(dataframe)
}
\arguments{
\item{dataframe}{dataframe}
}
\description{
This uses udpipe to tag the text. It then compresses all of the text 
so you have continuous POS tagging or the whole text. The udpipe package has to be
pre downloaded to run this.
}
\examples{
pp<-HistolMacDescrip(Mypath$Macroscopicdescription)
}
\keyword{Macroscopic}
