% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperEndoscopist}
\alias{EndoscChopperEndoscopist}
\title{EndoscChopperEndoscopist}
\usage{
EndoscChopperEndoscopist(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{The endoscopy text column}
}
\description{
This extracts the endoscopist from the report
}
\examples{
v<-TheOGDReportFinal
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
v<-EndoscChopperEndoscopist(Myendo,'Endoscopist')
}
\keyword{Endoscopist}
\keyword{extraction}
