% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperEndoscopist}
\alias{EndoscChopperEndoscopist}
\title{EndoscChopperEndoscopist}
\usage{
EndoscChopperEndoscopist(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{The endoscopy text column}
}
\description{
This extracts the endoscopist from the report
}
\examples{
v<-EndoscChopperEndoscopist(Myendo,'Endoscopist')
}
\keyword{Endoscopist}
\keyword{extraction}
