% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperEndoscopist}
\alias{EndoscChopperEndoscopist}
\title{Cleans endoscopist column if present}
\usage{
EndoscChopperEndoscopist(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{The endoscopy text column}
}
\description{
If an endoscopist column is part of the dataset once the extractor
function has been used this cleans the endoscopist column from the report.
It gets rid of titles
It gets rid of common entries that are not needed.
It should be used after the Extractor and the optional ChopperNewLines
has been used.
}
\examples{
v<-EndoscChopperEndoscopist(Myendo,'Endoscopist')
}
\keyword{Endoscopist}
\keyword{extraction}
