% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperFindings}
\alias{EndoscChopperFindings}
\title{EndoscChopperFindings}
\usage{
EndoscChopperFindings(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the Findings performed from the report
}
\examples{
v<-EndoscChopperFindings(Myendo,'Findings')
}
\keyword{Procedure}
