% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperIndications}
\alias{EndoscChopperIndications}
\title{EndoscChopperIndications}
\usage{
EndoscChopperIndications(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the Indication from the report
}
\examples{
EndoscChopperIndications(v,'Indications')
}
\keyword{Indications}
