% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperInstrument}
\alias{EndoscChopperInstrument}
\title{EndoscChopperInstrument}
\usage{
EndoscChopperInstrument(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans the Instrument from the report
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v<-EndoscChopperInstrument(v,"Instrument")
}
\keyword{Instrument}
