% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperInstrument}
\alias{EndoscChopperInstrument}
\title{Cleans instrument column if present}
\usage{
EndoscChopperInstrument(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This cleans Instument column from the report assuming such a column exists
(where instrument usually refers to the endoscope number being used.
It gets rid of common entries that are not needed.
It should be used after the Extractor and the optional
ChopperNewLines has been used.
}
\examples{
v<-EndoscChopperInstrument(Myendo,'Instrument')
}
\keyword{Instrument}
