% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscChopperMeds}
\alias{EndoscChopperMeds}
\title{Cleans medication column if present}
\usage{
EndoscChopperMeds(dataframe, MedColumn)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{MedColumn}{column of interest}
}
\description{
This cleans medication column from the report assuming such a column exists.
It gets rid of common entries that are not needed. It also splits the
medication into fentanyl and midazolam doses for use in the global rating
scale tables later. It should be used after the Extractor and the optional
ChopperNewLines has been used.
}
\examples{
v<-EndoscChopperMeds(Myendo,'Medications')
}
\keyword{Endoscopy}
\keyword{medications}
