% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{EndoscEndoscopist}
\alias{EndoscEndoscopist}
\title{Clean endoscopist column}
\usage{
EndoscEndoscopist(EndoscopistColumn)
}
\arguments{
\item{EndoscopistColumn}{The endoscopy text column}
}
\value{
This returns a character vector
}
\description{
If an endoscopist column is part of the dataset once the extractor
function has been used this cleans the endoscopist column from the report.
It gets rid of titles
It gets rid of common entries that are not needed.
It should be used after the textPrep function
}
\examples{
Myendo$Endoscopist <- EndoscEndoscopist(Myendo$Endoscopist)
}
\keyword{Endoscopist}
\keyword{extraction}
