% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscInstrument}
\alias{EndoscInstrument}
\title{Cleans instrument column if present}
\usage{
EndoscInstrument(EndoInstrument)
}
\arguments{
\item{EndoInstrument}{column of interest}
}
\value{
This returns a character vector
}
\description{
This cleans the Instument column from the report assuming such a column exists
(where instrument usually refers to the endoscope number being used.)
It gets rid of common entries that are not needed.
It should be used after the Extractor.
}
\examples{
Myendo$Instrument<-EndoscInstrument(Myendo$Instrument)
}
\keyword{Instrument}
