% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{EndoscInstrument}
\alias{EndoscInstrument}
\title{Clean instrument column}
\usage{
EndoscInstrument(EndoInstrument)
}
\arguments{
\item{EndoInstrument}{column of interest}
}
\value{
This returns a character vector
}
\description{
This cleans the Instument column from the report assuming such a column exists
(where instrument usually refers to the endoscope number being used.)
It gets rid of common entries that are not needed.
It should be used after the textPrep function.
Note this is possibly going to be deprecated in the next version 
as the endoscope coding used here is not widely used.
}
\examples{
Myendo$Instrument <- EndoscInstrument(Myendo$Instrument)
}
\seealso{
Other Endoscopy specific cleaning functions: \code{\link{EndoscEndoscopist}},
  \code{\link{EndoscMeds}}, \code{\link{EndoscopyEvent}}
}
\concept{Endoscopy specific cleaning functions}
\keyword{Instrument}
