% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscMeds}
\alias{EndoscMeds}
\title{Cleans medication column if present}
\usage{
EndoscMeds(MedColumn)
}
\arguments{
\item{MedColumn}{column of interest as a string vector}
}
\value{
This returns a dataframe
}
\description{
This cleans medication column from the report assuming such a column exists.
It gets rid of common entries that are not needed. It also splits the
medication into fentanyl and midazolam numeric doses for use. 
It should be used after the Extractor
}
\examples{
MyendoNew<-cbind(EndoscMeds(Myendo$Medications),Myendo)

}
\keyword{Endoscopy}
\keyword{medications}
