% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{Endoscopies}
\alias{Endoscopies}
\title{FakeEndoCreator}
\usage{
Endoscopies(x)
}
\arguments{
\item{x}{None needed}
}
\description{
Creates fake endoscopic data to play with as a spreadsheet format. It assumes
that some of the data has already been separated out. EndoRaw() for upper GI
and ColonEndoRaw() for lower GI are better functions for the real
 scenario of just getting the report as a series of unextracted text files.
(Their histology equivalents are pathRep and ColonpathRep respectively)
}
\examples{
Endoscopies(x)
}
\keyword{Fake}
\keyword{endoscopy}
