% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{EndoscopyEvent}
\alias{EndoscopyEvent}
\title{EndoscopyEvent}
\usage{
EndoscopyEvent(dataframe, EventColumn1, Procedure, Macroscopic, Histology)
}
\arguments{
\item{dataframe}{datafrane of interest}

\item{EventColumn1}{The relevant endoscopt free text column describing the findings}

\item{Procedure}{Column saying which procedure was performed}

\item{Macroscopic}{Column describing all the macroscopic specimens}

\item{Histology}{Column with free text histology (usually microscopic histology)}
}
\value{
This returns a character vector
}
\description{
This extracts the endoscopic event. It looks for the event term and then looks in the event sentence as well as the one above to see if
the location is listed. It needs to be run AFTER the HistolTypeAndSite function as emr needs to be
added to the event. Used in the OPCS4 coding
}
\examples{
# Myendo$EndoscopyEvent<-EndoscopyEvent(Myendo,"Findings",
#"ProcedurePerformed","MACROSCOPICALDESCRIPTION","HISTOLOGY")
}
\keyword{Find}
\keyword{and}
\keyword{replace}
