% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EndoscProcPerformed}
\alias{EndoscProcPerformed}
\title{Cleans Procedure performed column if present}
\usage{
EndoscProcPerformed(dataframe, ProcPerformed)
}
\arguments{
\item{dataframe}{dataframe with column of interest}

\item{ProcPerformed}{column of interest}
}
\description{
This cleans the Procedure Performed column from the report assuming
such a column exists. Procedure Performed relates to whether this was a
Gastroscopy or Colonoscopy and or the type of therapy used etc.
It gets rid of common entries that are not needed.
It should be used after the Extractor and the optional NewLines
has been used.
}
\examples{
v<-EndoscProcPerformed(Myendo,'ProcedurePerformed')
}
\keyword{Procedure}
