% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EntityPairs_OneSentence}
\alias{EntityPairs_OneSentence}
\title{EntityPairs_OneSentence}
\usage{
EntityPairs_OneSentence(dataframe, EventColumn)
}
\arguments{
\item{EventColumn1}{The relevant pathology text column}

\item{EventColumn2}{The alternative pathology text column}
}
\description{
This needs some blurb to be written. Used in the SentenceWordPairs
}
\examples{
# tbb<-EntityPairs_OneSentence(SelfOGD_Dunn,"MACROSCOPICALDESCRIPTION")
}
\keyword{PathPairLookup}
