% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{EntityPairs_OneSentence}
\alias{EntityPairs_OneSentence}
\title{EntityPairs_OneSentence}
\usage{
EntityPairs_OneSentence(inputText, list1, list2)
}
\arguments{
\item{inputText}{The relevant pathology text column}

\item{list1}{First list to refer to}

\item{list2}{The second list to look for}
}
\description{
Use to see if words from two lists co-exist within a sentence. Eg site and tissue type.
This function only looks in one sentence for the two terms. If you suspect the terms may
occur in adjacent sentences then use the EntityPairs_TwoSentence function.
}
\examples{
# tbb<-EntityPairs_OneSentence(Mypath$Histology,HistolType(),LocationList())
}
\keyword{PathPairLookup}
