% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{EntityPairs_TwoSentence}
\alias{EntityPairs_TwoSentence}
\title{EntityPairs_TwoSentence}
\usage{
EntityPairs_TwoSentence(inputString, list1, list2)
}
\arguments{
\item{EventColumn1}{The relevant pathology text column}

\item{EventColumn2}{The alternative pathology text column}
}
\description{
This is used to look for relationships between site and event especially for endoscopy events
}
\examples{
# tbb<-EntityPairs_TwoSentence(SelfOGD_Dunn,"FINDINGS")
}
\keyword{Find}
\keyword{and}
\keyword{replace}
