% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{Extractor}
\alias{Extractor}
\title{Extractor}
\usage{
Extractor(x, y, stra, strb, t)
}
\arguments{
\item{x}{the dataframe}

\item{y}{the column to extract from}

\item{stra}{the start of the boundary to extract}

\item{strb}{the end of the boundary to extract}

\item{t}{the column name to create}
}
\description{
This is the main extractor for the Endoscopy and Histology report
}
\examples{
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
}
\keyword{Extraction}
