% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{ExtrapolatefromDictionary}
\alias{ExtrapolatefromDictionary}
\title{Extrapolate from Dictionary}
\usage{
ExtrapolatefromDictionary(inputString, list)
}
\arguments{
\item{inputString}{The text string to process}

\item{list}{of words to iterate through}
}
\description{
Provides term mapping and extraction in one.
Standardises any term according to a mapping lexicon provided and then
extracts the term. This is
different to the DictionaryInPlaceReplace in that it provides a new column
with the extracted terms as opposed to changing it in place
}
\examples{
#Firstly we extract histology from the raw report
# The function then standardises the histology terms through a series of
# regular expressions and then extracts the type of tissue 
Mypath$Tissue<-suppressWarnings(
suppressMessages(
ExtrapolatefromDictionary(Mypath$Histology,HistolType()
)
)
)
rm(MypathExtraction)
}
\seealso{
Other Basic Column mutators: 
\code{\link{EntityPairs_OneSentence}()},
\code{\link{EntityPairs_TwoSentence}()},
\code{\link{ListLookup}()},
\code{\link{MyImgLibrary}()}
}
\concept{Basic Column mutators}
\keyword{Withdrawal}
