% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{GISymptomsList}
\alias{GISymptomsList}
\title{GISymptoms}
\usage{
GISymptomsList()
}
\description{
This function returns all the common GI symptoms. 
They are simply listed as is without grouping
or mapping. They have been derived from a manual 
list with synonyms derived from the UMLS Methatharus
using the browser.
}
\keyword{Event}
\keyword{extraction}
