% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{GRS_Type_Assess_By_Unit}
\alias{GRS_Type_Assess_By_Unit}
\title{GRS_Type_Assess_By_Unit
This extracts the polyps types from the data(for colonoscopy and flexible sigmoidosscopy data)}
\usage{
GRS_Type_Assess_By_Unit(x, ProcPerformed, Endo_Endoscopist, Dx, Histol)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column BxSize have been calculated}
}
\description{
GRS_Type_Assess_By_Unit
This extracts the polyps types from the data(for colonoscopy and flexible sigmoidosscopy data)
}
\keyword{Withdrawal}
