% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{GRS_Type_Assess_By_Unit}
\alias{GRS_Type_Assess_By_Unit}
\title{GRS_Type_Assess_By_Unit}
\usage{
GRS_Type_Assess_By_Unit(x, ProcPerformed, Endo_Endoscopist, Dx, Histol)
}
\arguments{
\item{x}{The dataframe}

\item{ProcPerformed}{The column containing the Procedure type performed}

\item{Endo_Endoscopist}{column containing the Endoscopist name}

\item{Dx}{The column with the Histological diagnosis}

\item{Histol}{The column with the Histology text in it}
}
\description{
This extracts the polyps types from the data (for colonoscopy and flexible sigmoidosscopy data)
and output the adenoma,adenocarcinoma and hyperplastic detection rate by endoscopist as well
as overall number of colonoscopies.
This will be extended to other GRS outputs in the future.
}
\examples{
v<-HistolChopperDx(Mypath,"Diagnosis")
v<-HistolChopperExtrapolDx(v,"Diagnosis")
v<-HistolChopperNumbOfBx(v,"Macroscopicdescription","specimen")
v<-HistolChopperBxSize(v,"Macroscopicdescription")
v<-Endomerge2(Myendo,"Dateofprocedure","HospitalNumber",v,"Dateofprocedure","HospitalNumber")
GRSTable<-GRS_Type_Assess_By_Unit(v,'ProcedurePerformed', 'Endoscopist','Diagnosis','Histology')
}
\keyword{Withdrawal}
