% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolAccessionNumber}
\alias{HistolAccessionNumber}
\title{Extract histological accession number}
\usage{
HistolAccessionNumber(dataframe, AccessionColumn, regString)
}
\arguments{
\item{dataframe}{dataframe name and}

\item{AccessionColumn}{the column name as a string.}

\item{regString}{regular expression needed as a string}
}
\description{
This extracts Accession Number data data from the report where one is
present. The Accession number relates to the actual specimen number as
ascribed by the pathology service.
}
\examples{
HistolAccessionNumber(Mypath,'Histology',
'SP-\\\\d{2}-\\\\d{7}')
}
\keyword{Accession}
\keyword{Sample}
\keyword{number}
