% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolAll}
\alias{HistolAll}
\title{Clean up all Histology data}
\usage{
HistolAll(dataframe)
}
\arguments{
\item{dataframe}{dataframe with column of interest}
}
\description{
This is a parent function for all the functions below
}
\examples{
mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
Mypath<-Extractor(Mypath,"PathReportWhole",mywords)
names(Mypath)<-c("Original","HospitalNumber","PatientName","DOB",
"GeneralPractitioner","Datereceived","ClinicalDetails",
"Macroscopicdescription","Histology","Diagnosis","HospitalNumber",
"PatientName","DOB","GeneralPractitioner","Dateofprocedure",
"ClinicalDetails","Macroscopicdescription","Histology","Diagnosis")
v<-HistolAll(Mypath)
rm(Mypath)
}
\keyword{Macroscopic}
