% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolBiopsyIndex}
\alias{HistolBiopsyIndex}
\title{HistolTissueIndex}
\usage{
HistolBiopsyIndex(dataframe, PathSite)
}
\arguments{
\item{PathSite}{The column that has the pathology locatio and tissue type from HistolTypeAndSite}
}
\description{
This returns a number for all the biopsies taken based on distance from orifice. It is for biopsies only
}
\examples{
# SelfOGD_Dunn<-read_excel("/home/rstudio/GenDev/DevFiles/EndoMineRFunctionDev/SelfOGD_Dunn.xlsx")
SelfOGD_Dunn$PathSite<-HistolTypeAndSite(SelfOGD_Dunn,"MACROSCOPICALDESCRIPTION","HISTOLOGY")
HistolBiopsyIndex(SelfOGD_Dunn,"PathSite") 
}
\keyword{Pathology}
\keyword{biopsy}
\keyword{index}
