% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{HistolBxSize}
\alias{HistolBxSize}
\title{Determine the largest biopsy size from the histology report}
\usage{
HistolBxSize(MacroColumn)
}
\arguments{
\item{MacroColumn}{Macdescrip}
}
\description{
This extracts the biopsy size from the report. If there are multiple
biopsies it will extract the overall size of each one (size is calculated
usually in cubic mm from the three dimensions provided). This will result
in row duplication.
}
\details{
This is usually from the Macroscopic description column.
}
\examples{
rr<-HistolBxSize(Mypath$Macroscopicdescription)
}
\keyword{biopsy}
\keyword{size}
