% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperAccessionNumber}
\alias{HistolChopperAccessionNumber}
\title{HistolChopperAccessionNumber}
\usage{
HistolChopperAccessionNumber(x, y, stra)
}
\arguments{
\item{x}{the dataframe name and}

\item{y}{the column name as a string.}

\item{stra}{regular expression needed as a string}
}
\description{
This extracts Accession Number data data from the report
}
\examples{
v<-Endomerge(PathDataFrameFinalColon,ColonFinal)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
Histoltree<-list("Clinical Details","Nature of specimen","Histology","Diagnosis","")
v<-ChopperNewLines(v,"Histo_ResultText")
for(i in 1:(length(Histoltree)-1)) {
 v<-Extractor(v,"Histo_ResultText",as.character(Histoltree[i]),as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
v<-HistolChopperAccessionNumber(v,"Histo_ResultText","SP-\\\\d{2}-\\\\d{7}")
}
\keyword{Accession}
\keyword{Sample}
\keyword{number}
