% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperAccessionNumber}
\alias{HistolChopperAccessionNumber}
\title{HistolChopperAccessionNumber}
\usage{
HistolChopperAccessionNumber(x, y, stra)
}
\arguments{
\item{x}{the dataframe name and}

\item{y}{the column name as a string.}

\item{stra}{regular expression needed as a string}
}
\description{
This extracts Accession Number data data from the report where one is present.
The Accession number relates to the actual specimen number as ascribed by the 
pathology service.
}
\examples{
v<-HistolChopperAccessionNumber(Mypath,"Histology","SP-\\\\d{2}-\\\\d{7}")
}
\keyword{Accession}
\keyword{Sample}
\keyword{number}
