% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperBxSize}
\alias{HistolChopperBxSize}
\title{HistolChopperBxSize}
\usage{
HistolChopperBxSize(x, y)
}
\arguments{
\item{x}{the dataframe}

\item{y}{Macdescrip}
}
\description{
This extracts the biopsy size from the report.This is usually from the Macroscopic description column.
}
\examples{
v<-HistolChopperBxSize(Mypath,'Macroscopicdescription')
}
\keyword{biopsy}
\keyword{size}
