% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperBxSize}
\alias{HistolChopperBxSize}
\title{HistolChopperBxSize}
\usage{
HistolChopperBxSize(x, y)
}
\arguments{
\item{x}{the dataframe}

\item{y}{Macdescrip}
}
\description{
This extracts the biopsy size from the report. If there are multiple 
biopsies it will extract the overall size of each one (size is calculated 
usually in cubic mm from the three dimensions provided). This will result
in row duplication.
}
\details{
This is usually from the Macroscopic description column.
}
\examples{
v<-HistolChopperBxSize(Mypath,'Natureofspecimen')
}
\keyword{biopsy}
\keyword{size}
