% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperDx}
\alias{HistolChopperDx}
\title{HistolChopperDx}
\usage{
HistolChopperDx(x, y)
}
\arguments{
\item{x}{the dataframe}

\item{y}{column containing the Hisopathology report}
}
\description{
This extracts Diagnosis data from the report. The Diagnosis is the overall 
impression of the pathologist for that specimen. At the moment, Only Capital 
D included (not lower case d) to make sure picks up subtitle header as 
opposed to mentioning 'diagnosis' as part of a sentence.  Column specific 
cleanup and negative remover have also been implemented here.
}
\examples{
v<-HistolChopperDx(Mypath,'Diagnosis')
}
\keyword{Diagnosis}
\keyword{Histology}
