% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperDx}
\alias{HistolChopperDx}
\title{HistolChopperDx}
\usage{
HistolChopperDx(x, y)
}
\arguments{
\item{x}{the dataframe}

\item{y}{column containing the Hisopathology report}
}
\description{
This extracts Diagnosis data from the report
}
\examples{
v<-PathDataFrameFinalColon
Histoltree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Clinical Details","Nature of specimen",
"Histology","Diagnosis","")
for(i in 1:(length(Histoltree)-1)) {
  Mypath<-Extractor(Mypath,"PathReportWhole",as.character(Histoltree[i]),
  as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
v<-HistolChopperDx(v,'Diagnosis')
}
\keyword{Diagnosis}
\keyword{Histology}
