% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperDx}
\alias{HistolChopperDx}
\title{HistolChopperDx}
\usage{
HistolChopperDx(x, y)
}
\arguments{
\item{x}{the dataframe}

\item{y}{column containing the Hisopathology report}
}
\description{
This extracts Diagnosis data from the report
}
\examples{
v<-HistolChopperDx(Mypath,'Diagnosis')
}
\keyword{Diagnosis}
\keyword{Histology}
