% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperExtrapolDx}
\alias{HistolChopperExtrapolDx}
\title{HistolChopperExtrapolDx}
\usage{
HistolChopperExtrapolDx(x, y)
}
\arguments{
\item{x}{the dataframe containing histology results,}

\item{y}{the column to extract dysplasia, cancer, and GIST from- often the Histology diagnosis column}
}
\description{
This extracts other specific diagnoses from the report
}
\examples{
Mypath<-PathDataFrameFinalColon
Histoltree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Clinical Details","Nature of specimen",
"Histology","Diagnosis","")
for(i in 1:(length(Histoltree)-1)) {
  Mypath<-Extractor(Mypath,"PathReportWhole",as.character(Histoltree[i]),
  as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
v<-HistolChopperExtrapolDx(Mypath,'Diagnosis')
}
\keyword{Histology}
\keyword{diagnosis}
