% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperExtrapolDx}
\alias{HistolChopperExtrapolDx}
\title{HistolChopperExtrapolDx}
\usage{
HistolChopperExtrapolDx(x, y)
}
\arguments{
\item{x}{the dataframe containing histology results,}

\item{y}{the column to extract dysplasia, cancer, and GIST from- often the Histology diagnosis column}
}
\description{
This extracts other specific diagnoses from the report
}
\examples{
HistolChopperExtrapolDx(v,'Diagnosis')
}
\keyword{Histology}
\keyword{diagnosis}
