% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperHistol}
\alias{HistolChopperHistol}
\title{HistolChopperHistol}
\usage{
HistolChopperHistol(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This extracts Histology details data from the report
}
\examples{
HistolChopperHistol(Mypath,'Histology')
}
\keyword{Histology}
