% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperHistol}
\alias{HistolChopperHistol}
\title{Extract the histology data from the report by removing negative findings}
\usage{
HistolChopperHistol(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This extracts Histology details data from the report. The Histology details
usually relate to the description of the histological report. This implements
 the negative remover and also adds further negative removing regexes. This 
may be refined in further iterations.
}
\examples{
t<-HistolChopperHistol(Mypath,'Histology')
}
\keyword{Histology}
