% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescrip}
\alias{HistolChopperMacDescrip}
\title{HistolChopperMacDescrip}
\usage{
HistolChopperMacDescrip(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{column to extract the numbers from}
}
\description{
This extracts numbers from written (spelt) numbers
}
\examples{
x <- HistolChopperMacDescrip(x, y)
}
\keyword{Macroscopic}
