% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescrip}
\alias{HistolChopperMacDescrip}
\title{Cleans spelt numbers in histology report}
\usage{
HistolChopperMacDescrip(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{column to extract the numbers from. Usually the column
with the Nature of the specimen or the Macroscopic description in it}
}
\description{
This extracts numbers from written (spelt) numbers in the Macroscopic 
description text. This means the text can then be used to extract the number 
and size of biopsies.This is used as part of the 
HistolChopperNumOfBx function below and normally not used as a stand alone 
function.
}
\examples{
t<-HistolChopperMacDescrip(Mypath, 'Macroscopicdescription')
}
\keyword{Macroscopic}
