% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescrip}
\alias{HistolChopperMacDescrip}
\title{HistolChopperMacDescrip}
\usage{
HistolChopperMacDescrip(x, y)
}
\arguments{
\item{x}{dataframe}

\item{y}{column to extract the numbers from. Usually the column
with the Nature of the specimen or the Macroscopic description in it}
}
\description{
This extracts numbers from written (spelt) numbers
}
\examples{
#' @examples v<-PathDataFrameFinalColon
Histoltree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Clinical Details","Nature of specimen",
"Histology","Diagnosis","")
for(i in 1:(length(Histoltree)-1)) {
  Mypath<-Extractor(Mypath,"PathReportWhole",as.character(Histoltree[i]),
  as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
HistolChopperMacDescrip(Mypath, "Natureofspecimen")
}
\keyword{Macroscopic}
