% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescripCleanup}
\alias{HistolChopperMacDescripCleanup}
\title{Clean up histological macroscopic description data}
\usage{
HistolChopperMacDescripCleanup(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest that describes the macroscopic specimen}
}
\description{
This extracts Macroscopic description data from the pathology report.
Macroscopic description usually relates to the number of specimens
retrieved, the size of each specimen and the location it was taken from.
The cleanup usually relates to the removal of top and tail characters such
as who reported the specimens etc.
}
\examples{
v<-HistolChopperMacDescripCleanup(Mypath,'Macroscopicdescription')
}
\keyword{Macroscopic}
