% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperMacDescripCleanup}
\alias{HistolChopperMacDescripCleanup}
\title{HistolChopperMacDescripCleanup}
\usage{
HistolChopperMacDescripCleanup(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This extracts Macroscopic description data from the report
}
\examples{
x <- HistolChopperMacDescripCleanup(x, y)
}
\keyword{Macroscopic}
