% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperNumbOfBx}
\alias{HistolChopperNumbOfBx}
\title{HistolChopperNumbOfBx}
\usage{
HistolChopperNumbOfBx(x, y, z)
}
\arguments{
\item{x}{the dataframe}

\item{y}{Macdescrip}
}
\description{
This extracts the number of biopsies taken from the pathology report. This is usually from the Macroscopic description column.
It collects everything from the regex [0-9]{1,2}.{0,3} to whatever the string boundary is (z)
}
\examples{
HistolChopperNumbOfBx(v,'Natureofspecimen','specimen')
}
\keyword{Biopsy}
\keyword{number}
