% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolChopperNumbOfBx}
\alias{HistolChopperNumbOfBx}
\title{HistolChopperNumbOfBx}
\usage{
HistolChopperNumbOfBx(x, y, z)
}
\arguments{
\item{x}{the dataframe}

\item{y}{Column containing the Macroscopic description text}

\item{z}{The keyword to remove and to stop at in the regex}
}
\description{
This extracts the number of biopsies taken from the pathology report. This is usually from the Macroscopic description column.
It collects everything from the regex [0-9]{1,2}.{0,3} to whatever the string boundary is (z)
}
\examples{
Mypath<-PathDataFrameFinalColon
Histoltree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Clinical Details","Nature of specimen",
"Histology","Diagnosis","")
for(i in 1:(length(Histoltree)-1)) {
  Mypath<-Extractor(Mypath,"PathReportWhole",as.character(Histoltree[i]),
  as.character(Histoltree[i+1]),gsub(" ","",as.character(Histoltree[i])))
}
v<-HistolChopperNumbOfBx(Mypath,'Natureofspecimen','specimen')
}
\keyword{Biopsy}
\keyword{number}
