% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolExtrapolDx}
\alias{HistolExtrapolDx}
\title{Extract specific diagnoses from the histology report}
\usage{
HistolExtrapolDx(dataframe, Column)
}
\arguments{
\item{dataframe}{dataframe containing histology results,}

\item{Column}{the column to extract dysplasia, cancer, and GIST from- often the
Histology diagnosis column}
}
\description{
This extracts other specific diagnoses from the report. These have been hard
coded to look for dysplasia cancer and GIST. Optional use.
}
\examples{
v<-HistolExtrapolDx(Mypath,'Diagnosis')
}
\keyword{Histology}
\keyword{diagnosis}
