% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{HistolNumbOfBx}
\alias{HistolNumbOfBx}
\title{Extract the number of biopsies taken from the histology report}
\usage{
HistolNumbOfBx(inputString, regString)
}
\arguments{
\item{inputString}{The input text to process}

\item{regString}{The keyword to remove and to stop at in the regex}
}
\description{
This extracts the number of biopsies taken from the pathology report.
This is usually from the Macroscopic description column.
It collects everything from the regex [0-9]{1,2}.{0,3}
to whatever the string boundary is (z).
}
\examples{
qq <- HistolNumbOfBx(Mypath$Macroscopicdescription, "specimen")
}
\seealso{
Other Histology specific cleaning functions: \code{\link{HistolBxSize}},
  \code{\link{HistolTypeAndSite}}
}
\concept{Histology specific cleaning functions}
\keyword{Biopsy}
\keyword{number}
