% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{HistolType}
\alias{HistolType}
\title{Use list of pathology types}
\usage{
HistolType()
}
\description{
This standardizes terms to describe the pathology tissue type being examined
}
\seealso{
Other NLP - Lexicons: 
\code{\link{BiopsyIndex}()},
\code{\link{EventList}()},
\code{\link{GISymptomsList}()},
\code{\link{LocationListLower}()},
\code{\link{LocationListUniversal}()},
\code{\link{LocationListUpper}()},
\code{\link{LocationList}()},
\code{\link{RFACath}()},
\code{\link{WordsToNumbers}()}
}
\concept{NLP - Lexicons}
\keyword{Pathology}
\keyword{type}
