% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{HistolType}
\alias{HistolType}
\title{Extract pathology type}
\usage{
HistolType()
}
\description{
This standardizes terms to describe the pathology tissue type being examined
}
\keyword{Pathology}
\keyword{type}
