% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{HistolTypeAndSite}
\alias{HistolTypeAndSite}
\title{Extract the site a specimen was removed from as well as the type}
\usage{
HistolTypeAndSite(inputString1, inputString2, procedureString)
}
\arguments{
\item{inputString1}{The first column to look in}

\item{inputString2}{The second column to look in}

\item{procedureString}{The column with the procedure in it}
}
\value{
a list with two columns, one is the type and site and the other
is the index to be used for OPCS4 coding later if needed.
}
\description{
This needs some blurb to be written. Used in the OPCS4 coding
}
\examples{
Myendo2<-Endomerge2(Myendo,'Dateofprocedure','HospitalNumber',
Mypath,'Dateofprocedure','HospitalNumber')
PathSiteAndType <- HistolTypeAndSite(Myendo2$PathReportWhole,
Myendo2$Macroscopicdescription, Myendo2$ProcedurePerformed)
}
\seealso{
Other Histology specific cleaning functions: \code{\link{HistolBxSize}},
  \code{\link{HistolNumbOfBx}}
}
\concept{Histology specific cleaning functions}
\keyword{Find}
\keyword{and}
\keyword{replace}
