% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_EndoPathExtractors.R
\name{HistolTypeAndSite}
\alias{HistolTypeAndSite}
\title{HistolTypeAndSite}
\usage{
HistolTypeAndSite(inputString1, inputString2, procedureString)
}
\arguments{
\item{inputString1}{The first column to look in}

\item{inputString2}{The second column to look in}

\item{procedureString}{The column with the procedure in it}
}
\value{
a list with two columns, one is the type and site and the other
is the index to be used for OPCS4 coding later if needed.
}
\description{
This needs some blurb to be written. Used in the OPCS4 coding
}
\examples{
Myendo$PathSite <- HistolTypeAndSite(
  Myendo$Original.x,
  Myendo$mscroscopicaldescription, Myendo$procedureperformed
) # This example needs correction
}
\keyword{Find}
\keyword{and}
\keyword{replace}
