% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{HistolTypeAndSite}
\alias{HistolTypeAndSite}
\title{HistolTypeAndSite}
\usage{
HistolTypeAndSite(inputString1, inputString2, procedureString)
}
\arguments{
\item{Procedure}{The procedure performed}

\item{EventColumn1}{The relevant pathology text column}

\item{EventColumn2}{The alternative pathology text column}
}
\value{
a list with two columns, one is the type and site and the other
is the index to be used for OPCS4 coding later if needed.
}
\description{
This needs some blurb to be written. Used in the OPCS4 coding
}
\examples{
# SelfOGD_Dunn$PathSite<-HistolTypeAndSite(SelfOGD_Dunn$PROCEDUREPERFORMED,SelfOGD_Dunn$MACROSCOPICALDESCRIPTION,SelfOGD_Dunn$HISTOLOGY)
}
\keyword{Find}
\keyword{and}
\keyword{replace}
