% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoFakeData.R
\name{Histop_df}
\alias{Histop_df}
\title{FakeHistolCreator}
\usage{
Histop_df(x)
}
\arguments{
\item{x}{None needed}
}
\description{
Creates fake histology data to play with
}
\examples{
Histop_df()
}
\keyword{Fake}
\keyword{data}
\keyword{histology}
\keyword{spreadshet}
