% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{HowManyTests}
\alias{HowManyTests}
\title{HowManyTests}
\usage{
HowManyTests(x, Indication, Endo_ResultPerformed, StringToSearch)
}
\arguments{
\item{x}{dataframe}

\item{Indication}{Indication column}

\item{Endo_ResultPerformed}{column containing date the Endoscopy was performed}

\item{StringToSearch}{The string in the Indication to search for}
}
\description{
Get an overall idea of how many endoscopies have been done for an indication by year and month
}
\examples{
v<-TheOGDReportFinal
Myendo<-TheOGDReportFinal
Myendo$OGDReportWhole<-gsub("2nd Endoscopist:","Second endoscopist:",Myendo$OGDReportWhole)
EndoscTree<-list("Hospital Number:","Patient Name:","General Practitioner:",
"Date of procedure:","Endoscopist:","Second endoscopist:","Medications",
"Instrument","Extent of Exam:","Indications:","Procedure Performed:","Findings:",
"Endoscopic Diagnosis:")
for(i in 1:(length(EndoscTree)-1)) {
 Myendo<-Extractor(Myendo,"OGDReportWhole",as.character(EndoscTree[i]),
 as.character(EndoscTree[i+1]),as.character(EndoscTree[i]))
}
Myendo$Dateofprocedure<-as.Date(Myendo$Dateofprocedure)
HowManyTests(Myendo,'Indications','Dateofprocedure','Surv')
}
\keyword{Tests}
\keyword{number}
