% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{HowManyTests}
\alias{HowManyTests}
\title{HowManyTests}
\usage{
HowManyTests(x, Indication, Endo_ResultPerformed, StringToSearch)
}
\arguments{
\item{x}{dataframe}

\item{Indication}{Indication column}

\item{Endo_ResultPerformed}{column containing date the Endoscopy was performed}

\item{StringToSearch}{The string in the Indication to search for}
}
\description{
Get an overall idea of how many endoscopies have been done for an indication by year and month
}
\examples{
HowManyTests(v,'EXAMINATION','Endo_ResultPerformed','Abdo')
}
\keyword{Tests}
\keyword{number}
