% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{HowManyTests}
\alias{HowManyTests}
\title{HowManyTests}
\usage{
HowManyTests(x, Indication, Endo_ResultPerformed, StringToSearch)
}
\arguments{
\item{x}{dataframe}

\item{Indication}{Indication column}

\item{Endo_ResultPerformed}{column containing date the Endoscopy was performed}

\item{StringToSearch}{The string in the Indication to search for}
}
\description{
Get an overall idea of how many endoscopies have been done for an indication by year and month. This is a 
more involved version of SurveillanceCapacity function. It takes string for the Indication for the test
}
\details{
This returns a list which contains a plot (number of tests for that indication over time
and a table with the same information broken down by month and year).
}
\examples{
HowManyTests(Myendo,'Indications','Dateofprocedure','Surv')
}
\keyword{Tests}
\keyword{number}
