% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_IBD.R
\name{IBD_Scores}
\alias{IBD_Scores}
\title{Cleans medication column if present}
\usage{
IBD_Scores(inputColumn1)
}
\arguments{
\item{inputColumn1}{column of interest as a string vector}
}
\value{
This returns a dataframe with all the scores in it
}
\description{
This extracts all of the relevant IBD scores where present
from the medical text.
}
\examples{
 df<-read.csv("/Users/sebastianzeki/Documents/AnonKCHendo6m.csv",stringsAsFactors = FALSE)
IBD_Scores(df$ER_FINDINGS_STR)
MydfNew <- cbind(IBD_Scores(df$ER_FINDINGS_STR), df)
}
\keyword{IBD}
\keyword{scores}
