% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{ListLookup}
\alias{ListLookup}
\title{ListLookup functions:}
\usage{
ListLookup(theframe, y, myNotableWords)
}
\arguments{
\item{theframe}{the dataframe,}

\item{y}{the column of interest,}

\item{myNotableWord}{list of wirds you are interested in}
}
\description{
Input is a dataframe with x=dataframe name, y=column name you are doing text mining from and PropThreshold= the number of rows that have the terms you are interested in. theframe is the dataframe, y is the columnn of interest and PropThreshold is the Proportion of reports Threshold for the graph. The aim here is simply to produce a document term matrix to get the frequency of all the words, then extract the words you are interested in with tofind (I think) then find which reports have those words. Then find what proportion of the reports have those terms
}
\examples{
myNotableWords<-c('arrett','oeliac')
 ListLookup(v,'Findings',myNotableWords)
}
\keyword{Lookup}
