% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{ListLookup}
\alias{ListLookup}
\title{Extract from report, using words from a list}
\usage{
ListLookup(theframe, EndoReportColumn, myNotableWords)
}
\arguments{
\item{theframe}{the dataframe,}

\item{EndoReportColumn}{the column of interest,}

\item{myNotableWords}{list of words you are interested in}
}
\description{
The aim here is simply to
produce a document term matrix to get the frequency
of all the words, then extract the words you are
interested in with tofind then find which reports
have those words. Then find what proportion of the reports
have those terms.
}
\examples{
#The function relies on defined a list of
# words you are interested in and then choosing the column you are
# interested in looking in for these words. This can be for histopathology
# free text columns or endoscopic. In this example it is for endoscopic
# columns
myNotableWords<-c('arrett','oeliac')
ListLookup(Myendo,'Findings',myNotableWords)
}
\keyword{Lookup}
