% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{ListLookup}
\alias{ListLookup}
\title{ListLookup}
\usage{
ListLookup(theframe, y, myNotableWords)
}
\arguments{
\item{theframe}{the dataframe,}

\item{y}{the column of interest,}

\item{myNotableWords}{list of words you are interested in}
}
\description{
The aim here is simply to 
produce a document term matrix to get the frequency 
of all the words, then extract the words you are 
interested in with tofind then find which reports 
have those words. Then find what proportion of the reports 
have those terms.
}
\examples{
myNotableWords<-c('arrett','oeliac')
tt<-ListLookup(Myendo,'Findings',myNotableWords)
}
\keyword{Lookup}
