% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{LocationList}
\alias{LocationList}
\title{Use list of upper and lower GI standard locations}
\usage{
LocationList()
}
\description{
The is a list of standard locations at endoscopy. It used for the site of biopsies/EMRs 
and potentially in functions looking at the site of a 
therapeutic event. It just returns the list in the function.
}
\seealso{
Other NLP - Lexicons: 
\code{\link{BiopsyIndex}()},
\code{\link{EventList}()},
\code{\link{GISymptomsList}()},
\code{\link{HistolType}()},
\code{\link{LocationListLower}()},
\code{\link{LocationListUniversal}()},
\code{\link{LocationListUpper}()},
\code{\link{RFACath}()},
\code{\link{WordsToNumbers}()}
}
\concept{NLP - Lexicons}
\keyword{Location}
