% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{MetricByEndoscopist}
\alias{MetricByEndoscopist}
\title{MetricByEndoscopist}
\usage{
MetricByEndoscopist(x, y, z)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column (numeric data) of interest}

\item{z}{The endoscopist column}
}
\description{
This takes any numerical metric in the dataset and plots it by endoscopist.
It of course relies on a Endoscopist column being present
}
\examples{
Myendo<-EndoscChopperMeds(Myendo,'Medications')
Fent<-MetricByEndoscopist(Myendo,'Endoscopist','Fent')
}
\keyword{Endoscopist}
