% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{MetricByEndoscopist}
\alias{MetricByEndoscopist}
\title{Number of biopsies
This assess how many biopsies have been taken}
\usage{
MetricByEndoscopist(x, y, z)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column (numeric data) of interest}

\item{z}{The endoscopist column}
}
\description{
Number of biopsies
This assess how many biopsies have been taken
}
\keyword{Endoscopist}
