% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{MetricByEndoscopist}
\alias{MetricByEndoscopist}
\title{Plot a metric by endoscopist}
\usage{
MetricByEndoscopist(dataframe, Column, EndoscopistColumn)
}
\arguments{
\item{dataframe}{The dataframe}

\item{Column}{The column (numeric data) of interest}

\item{EndoscopistColumn}{The endoscopist column}
}
\description{
This takes any of the numerical metrics in the dataset and plots it by
endoscopist.
It of course relies on a Endoscopist column being present
}
\examples{
#The function gives a table with any numeric
# metric by endoscopist
# In this example we tabulate medication by
# endoscopist
# Lets bind the output of EndoscMeds to the main dataframe so we
# have a complete dataframe with all the meds extracted
MyendoNew<-cbind(EndoscMeds(Myendo$Medications),Myendo)

# Now lets look at the fentanly use per Endoscopist:
kk<-MetricByEndoscopist(MyendoNew,'Endoscopist','Fent')
rm(Myendo)
}
\keyword{Endoscopist}
