% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{MetricByEndoscopist}
\alias{MetricByEndoscopist}
\title{Number of biopsies
This assess how many biopsies have been taken}
\usage{
MetricByEndoscopist(x, y, z)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column (numeric data) of interest}

\item{z}{The endoscopist column}
}
\description{
Number of biopsies
This assess how many biopsies have been taken
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v<-BarrettsDataAccord_Prague(v,"Endo_ResultText")
v<-EndoscChopperMeds(v,"Medications")
Fent<-MetricByEndoscopist(v,"Endoscopist","Fent")
}
\keyword{Endoscopist}
