% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{MetricByEndoscopist}
\alias{MetricByEndoscopist}
\title{Plot a metric by endoscopist}
\usage{
MetricByEndoscopist(dataframe, Column, EndoscopistColumn)
}
\arguments{
\item{dataframe}{The dataframe}

\item{Column}{The column (numeric data) of interest}

\item{EndoscopistColumn}{The endoscopist column}
}
\description{
This takes any numerical metric in the dataset and plots it by endoscopist.
It of course relies on a Endoscopist column being present
}
\examples{
#The function plots any numeric metric by endoscopist
# and also gives a table with it. In this example we plot medication by
# endoscopist
Myendo<-EndoscMeds(Myendo,'Medications')
MetricByEndoscopist(Myendo,'Endoscopist','Fent')
rm(Myendo)
}
\keyword{Endoscopist}
