% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{NegativeRemove}
\alias{NegativeRemove}
\title{Remove negative and normal sentences}
\usage{
NegativeRemove(inputText)
}
\arguments{
\item{inputText}{column of interest}
}
\value{
This returns a column within a dataframe. THis should be changed to a 
character vector eventually
}
\description{
Extraction of the negative sentences so that normal findings can be
removed and not counted when searching for true diseases. eg remove
'No evidence of candidal infection' so it doesn't get included if
looking for candidal infections. It is used by default as part of
the textPrep function but can be turned off as an optional parameter
}
\examples{
# Build a character vector and then
# incorporate into a dataframe
anexample<-c("There is no evidence of polyp here",
"Although the prep was poor,there was no adenoma found",
"The colon was basically inflammed, but no polyp was seen",
"The Barrett's segment was not biopsied",
"The C0M7 stretch of Barrett's was flat")
anexample<-data.frame(anexample)
names(anexample)<-"Thecol"
# Run the function on the dataframe and it should get rid of sentences (and
# parts of sentences) with negative parts in them.
hh<-NegativeRemove(anexample$Thecol)
}
\seealso{
Other NLP - Text Cleaning and Extraction: 
\code{\link{ColumnCleanUp}()},
\code{\link{DictionaryInPlaceReplace}()},
\code{\link{Extractor}()},
\code{\link{textPrep}()}
}
\concept{NLP - Text Cleaning and Extraction}
\keyword{Negative}
\keyword{Sentences}
