% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanUp.R
\name{NegativeRemove}
\alias{NegativeRemove}
\title{NegativeRemove}
\usage{
NegativeRemove(x, y)
}
\arguments{
\item{x}{dataframe with column of interest}

\item{y}{column of interest}
}
\description{
This extracts removes negative diagnoses from the report
}
\keyword{Negative}
\keyword{Sentences}
