% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{NumberPerformed}
\alias{NumberPerformed}
\title{Determine overall number of procedures performed}
\usage{
NumberPerformed(dataframe, EndoscopistColumn, IndicationColumn)
}
\arguments{
\item{dataframe}{The dataframe}

\item{EndoscopistColumn}{The column containing the Endoscopists names}

\item{IndicationColumn}{The column containing the Indication for the examination}
}
\description{
Determines the number of endoscopies done by an endoscopist
by type of endosopy and indication for a given timeframe
As per BSG recommendations for Upper GI minimum number of
gastroscopies in a year (although here the time frame is user
defined)
}
\examples{
Myendo<-Myendo[grepl('Gastroscopy',Myendo$ProcedurePerformed),]
 NumberPerformed(Myendo,'Endoscopist','Indications')
}
\keyword{Withdrawal}
