% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PatientFlow_CircosPlots}
\alias{PatientFlow_CircosPlots}
\title{Circos plot functions:}
\usage{
PatientFlow_CircosPlots(x, Endo_ResultPerformed, HospNum_Id, ProcPerformed)
}
\arguments{
\item{x}{dataframe}

\item{Endo_ResultPerformed}{the column containing the date of the procedure}

\item{ProcPerformed}{The procedure that you want to plot (eg EMR,
radiofrequency ablation for Barrett's but can be any dscription of a procedure you desire)}

\item{Indication}{the column containing the indication for the procedure}
}
\description{
This allows us to look at the overall flow from one type of procedure to another using circos plots
}
\examples{
v<-Endomerge2(PathDataFrameFinal,OGDReportWhole)
names(v)<-c("Endo_ResultText","Endo_ResultPerformed","HospNumId","Histo_ResultText","Histo_ResultPerformed","HospNumId2","Days")
EndoscTree<-list('Hospital Number','Patient Name','General Practitioner','Endoscopist',
'nd Endoscopist','Medications','Instrument','Extent of Exam','Indications','Procedure Performed',
'Findings','Endoscopic Diagnosis','')
for(i in 1:(length(EndoscTree)-1)) {
  v<-Extractor(v,'Endo_ResultText',as.character(EndoscTree[i]),as.character(EndoscTree[i+1]),
  as.character(EndoscTree[i]))
}
v<-BarrettsDataAccord_Prague(v,"Endo_ResultText")
}
\keyword{Circos}
