% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PatientFlow_CircosPlots}
\alias{PatientFlow_CircosPlots}
\title{PatientFlow_CircosPlots}
\usage{
PatientFlow_CircosPlots(x, Endo_ResultPerformed, HospNum_Id, ProcPerformed)
}
\arguments{
\item{x}{dataframe}

\item{Endo_ResultPerformed}{the column containing the date of the procedure}

\item{HospNum_Id}{Column with the patient's unique hospital number}

\item{ProcPerformed}{The procedure that you want to plot (eg EMR,
radiofrequency ablation for Barrett's but can be 
any dscription of a procedure you desire)}
}
\description{
This allows us to look at the overall flow from one 
type of procedure to another using circos plots.
}
\keyword{Circos}
