% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypLocator}
\alias{PolypLocator}
\title{PolypLocator}
\usage{
PolypLocator(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column containing the SampleLocation from the TermStandardLocation}
}
\description{
This should be used after the TermStandardizer
as it relies on the presence of a SampleLocation column
which the TermStandardLocation produces. 
It should be used after the PolypTidyUpLocator
}
\examples{
f<-TermStandardLocation(Mypath,'Histology')
f<-PolypLocator(f,'SampleLocation')
}
\keyword{Withdrawal}
