% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypLocator}
\alias{PolypLocator}
\title{Standardises the location of polyps
This assess the size of the biopies take}
\usage{
PolypLocator(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column containing the SampleLocation from the Term Standardizer}
}
\description{
Standardises the location of polyps
This assess the size of the biopies take
}
\keyword{Withdrawal}
