% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypLocator}
\alias{PolypLocator}
\title{Determine polyp location}
\usage{
PolypLocator(dataframe, SampleLocationColumn)
}
\arguments{
\item{dataframe}{The dataframe}

\item{SampleLocationColumn}{The column containing the SampleLocation from the
TermStandardLocation}
}
\description{
This should be used after the TermStandardizer
as it relies on the presence of a SampleLocation column
which the TermStandardLocation produces.
}
\examples{
mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
Mypath<-Extractor(Mypath,"PathReportWhole",mywords)
names(Mypath)[names(Mypath) == 'Datereceived'] <- 'Dateofprocedure'
Mypath$Dateofprocedure <- as.Date(Mypath$Dateofprocedure)
# The polyp locator then determines where the biopsies were taken from
# by assessing the SampleLocation column which comes from the
# TermStandardLocation function.
f<-TermStandardLocation(Mypath,'Histology')
mm<-PolypLocator(f,'SampleLocation')
rm(Mypath)
}
\keyword{Withdrawal}
