% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypLocator}
\alias{PolypLocator}
\title{Standardises the location of polyps
This assess the size of the biopies take}
\usage{
PolypLocator(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column containing the SampleLocation from the Term Standardizer}
}
\description{
Standardises the location of polyps
This assess the size of the biopies take
}
\examples{
v<-PathDataFrameFinalColon
Histoltree<-list('Clinical Details','Nature of specimen','Histology','Diagnosis','')
for(i in 1:(length(Histoltree)-1)) {
 v<-Extractor(v,'PathReportWhole',as.character(Histoltree[i]),
 as.character(Histoltree[i+1]),gsub(' ','',as.character(Histoltree[i])))
}
f<-TermStandardLocation(v,'Histology')
f<-PolypLocator(f,'SampleLocation')
}
\keyword{Withdrawal}
