% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypLocator}
\alias{PolypLocator}
\title{Determine polyp location}
\usage{
PolypLocator(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{The column containing the SampleLocation from the 
TermStandardLocation}
}
\description{
This should be used after the TermStandardizer
as it relies on the presence of a SampleLocation column
which the TermStandardLocation produces.
It should be used after the PolypTidyUpLocator
}
\examples{
Histoltree <-list("Hospital Number:","Patient Name:",
"General Practitioner:","Date received:","Clinical Details",
 "Nature of specimen","Histology","Diagnosis",""
)

for (i in 1:(length(Histoltree) - 1)) {
 Mypath <-Extractor(
     Mypath, "PathReportWhole",as.character(Histoltree[i]),
     as.character(Histoltree[i + 1]),
     gsub(" ", "", as.character(Histoltree[i]))
   )
}
names(Mypath)[names(Mypath) == 'Datereceived'] <- 'Dateofprocedure'
Mypath$Dateofprocedure <- as.Date(Mypath$Dateofprocedure)
# The polyp locator then determines where the biopsies were taken from
# by assessing the SampleLocation column which comes from the
# TermStandardLocation function.
f<-TermStandardLocation(Mypath,'Histology')
f<-PolypLocator(f,'SampleLocation')
}
\keyword{Withdrawal}
