% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypTidyUpLocator}
\alias{PolypTidyUpLocator}
\title{Standardises the location of polyps
This assess the size of the biopies take}
\usage{
PolypTidyUpLocator(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{The column containing the SampleLocation from the Term Standardizer}
}
\description{
Standardises the location of polyps
This assess the size of the biopies take
}
\examples{
f<-TermStandardLocation(Mypath,'Histology')
}
\keyword{Withdrawal}
