% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypTidyUpLocator}
\alias{PolypTidyUpLocator}
\title{Clean the polyp location}
\usage{
PolypTidyUpLocator(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{The column containing the
SampleLocation from the Term Standardizer}
}
\description{
This cleans up the polyps from the TermStandardLocation function
}
\examples{
#' # The polyp locator then determines where the biopsies were taken from
# by assessing the SampleLocation column which comes from the
# TermStandardLocation function. This should also be run before the polyp 
# locator function
f<-TermStandardLocation(Mypath,'Histology')
f<-PolypTidyUpLocator(f,'SampleLocation')
}
\keyword{Withdrawal}
