% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{PolypTidyUpLocator}
\alias{PolypTidyUpLocator}
\title{PolypTidyUpLocator}
\usage{
PolypTidyUpLocator(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{The column containing the 
SampleLocation from the Term Standardizer}
}
\description{
This cleans up the polyps from the TermStandardLocation function
}
\examples{
f<-TermStandardLocation(Mypath,'Histology')
f<-PolypTidyUpLocator(f,'SampleLocation')
}
\keyword{Withdrawal}
