% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{RFACath}
\alias{RFACath}
\title{Use list of catheters used in radiofrequency ablation}
\usage{
RFACath()
}
\description{
The takes a list of catheters used in radiofrequency ablation
}
\examples{
# No example needed
}
\keyword{RFA}
