% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SampleLocator}
\alias{SampleLocator}
\title{Locate where samples are taken from}
\usage{
SampleLocator(dataframe, SampleLocationColumn)
}
\arguments{
\item{dataframe}{The dataframe}

\item{SampleLocationColumn}{SampleLocation from the TermStandardizer}
}
\description{
This assess where samples are taken from.
This should be used after the TermStandardLocation
as it relies on the presence of a SampleLocation column
which the TermStandardLocation produces.
It can be used after the PolypTidyUpLocator
although you could just run it after the TermStandardLocation
to get an overview of where samples were taken from.
It will tell you the sites sampled without duplication
}
\examples{
#Firstly we extract histology from the raw report
# using the extractor function
Histoltree <-list("Hospital Number:","Patient Name:",
"General Practitioner:","Date received:","Clinical Details",
 "Nature of specimen","Histology","Diagnosis",""
)

for (i in 1:(length(Histoltree) - 1)) {
 Mypath <-Extractor(
     Mypath, "PathReportWhole",as.character(Histoltree[i]),
     as.character(Histoltree[i + 1]),
     gsub(" ", "", as.character(Histoltree[i]))
   )
}
names(Mypath)[names(Mypath) == 'Datereceived'] <- 'Dateofprocedure'
Mypath$Dateofprocedure <- as.Date(Mypath$Dateofprocedure)
# The function needs all the terms to be standardised first so the
# function TermStandardLocation needs to be run first
f<-TermStandardLocation(Mypath,'Histology')
# The sample locator then determines where the biopsies were taken from
# by assessing the SampleLocation column which comes from the
# TermStandardLocation function.
f<-SampleLocator(f,'SampleLocation')
rm(Mypath)
}
\keyword{Sample}
\keyword{location}
