% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SampleLocator}
\alias{SampleLocator}
\title{Determines the site of samples
This assess where samples are taken from. This should be used after the TermStandardizer and can be used after the PolypTidyUpLocator 
although you could just run it after the term stanardiser to get an overview of where samples were taken from. It will tell you the
sites sampled without duplication}
\usage{
SampleLocator(x, y)
}
\arguments{
\item{x}{The dataframe}

\item{y}{SampleLocation from the TermStandardizer}
}
\description{
Determines the site of samples
This assess where samples are taken from. This should be used after the TermStandardizer and can be used after the PolypTidyUpLocator 
although you could just run it after the term stanardiser to get an overview of where samples were taken from. It will tell you the
sites sampled without duplication
}
\examples{
v<-PathDataFrameFinalColon
Histoltree<-list('Clinical Details','Nature of specimen','Histology','Diagnosis','')
for(i in 1:(length(Histoltree)-1)) {
 v<-Extractor(v,'PathReportWhole',as.character(Histoltree[i]),
 as.character(Histoltree[i+1]),gsub(' ','',as.character(Histoltree[i])))
}
f<-TermStandardLocation(v,'Histology')
f<-SampleLocator(f,'SampleLocation')
}
\keyword{Sample}
\keyword{location}
