% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Sets the publication theme for all the ggplots}
\usage{
scale_colour_Publication(...)
}
\description{
Sets the publication theme for all the ggplots
}
\examples{
# None needed
}
\keyword{ggplot}
\keyword{themes}
