% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Set the fills for all the ggplots}
\usage{
scale_fill_Publication()
}
\description{
This standardises the fills for any ggplot plot produced.
If you do use it, like all ggplots it can be extended using the 
"+" to add whatever else is necessary
}
\examples{
# None needed
}
\seealso{
Other Data Presentation helpers: 
\code{\link{EndoBasicGraph}()},
\code{\link{scale_colour_Publication}()},
\code{\link{theme_Publication}()}
}
\concept{Data Presentation helpers}
\keyword{ggplot}
\keyword{themes}
