% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveilFirstTest}
\alias{SurveilFirstTest}
\title{Extracts the first test only per patient}
\usage{
SurveilFirstTest(dataframe, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
Extracts the first test only per patient and returns a new dataframe listing the
patientID and the first test done
}
\examples{
dd <- SurveilFirstTest(
  Myendo, "HospitalNumber",
  "Dateofprocedure"
)
}
\seealso{
Other Basic Analysis - Surveillance Functions: \code{\link{HowManyOverTime}},
  \code{\link{SurveilLastTest}},
  \code{\link{SurveilTimeByRow}},
  \code{\link{TimeToStatus}}
}
\concept{Basic Analysis - Surveillance Functions}
\keyword{Surveillance}
