% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceCapacity}
\alias{SurveillanceCapacity}
\title{Number of tests done per month}
\usage{
SurveillanceCapacity(dataframe, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{dataframe}

\item{Endo_ResultPerformed}{Column with the date the Endoscopy was performed}
}
\description{
This determines the number of tests done per month
}
\examples{
em<-SurveillanceCapacity(Myendo,'Dateofprocedure')
}
\keyword{cats}
