% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceFirstTest}
\alias{SurveillanceFirstTest}
\title{Extracts the first test only per patient}
\usage{
SurveillanceFirstTest(dataframe, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
Extracts the first test only per patient
}
\examples{
em<-SurveillanceFirstTest(Myendo,'HospitalNumber',
'Dateofprocedure')
}
\keyword{Surveillance}
