% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceFirstTest}
\alias{SurveillanceFirstTest}
\title{SurveillanceFirstTest}
\usage{
SurveillanceFirstTest(x, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{x}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
Extract the first test only per patient
}
\examples{
SurveillanceFirstTest(v,'PatientID','Endo_ResultPerformed')
}
\keyword{Surveillance}
