% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceLastTest}
\alias{SurveillanceLastTest}
\title{Extract the last test done by a patient only}
\usage{
SurveillanceLastTest(dataframe, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{dataframe}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
Extracts the last test only per patient
}
\examples{
em<-SurveillanceLastTest(Myendo,'HospitalNumber','Dateofprocedure')
}
\keyword{Surveillance}
