% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceLastTest}
\alias{SurveillanceLastTest}
\title{SurveillanceLastTest}
\usage{
SurveillanceLastTest(x, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{x}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
Extract the last test only per patient
}
\examples{
em<-SurveillanceLastTest(Myendo,'HospitalNumber','Dateofprocedure')
}
\keyword{Surveillance}
