% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceLastToNow}
\alias{SurveillanceLastToNow}
\title{Extract the last test done by a patient in days and how long ago}
\usage{
SurveillanceLastToNow(x, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{x}{dataframe}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
This determines the last test done by that patient and the time
between now and that last test in days
}
\examples{
em<-SurveillanceLastToNow(Myendo,'HospitalNumber',
'Dateofprocedure')
}
\keyword{Surveillance}
