% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveillanceTimeByRow}
\alias{SurveillanceTimeByRow}
\title{SurveillanceTimeByRow}
\usage{
SurveillanceTimeByRow(x, HospNum_Id, Endo_ResultPerformed)
}
\arguments{
\item{x}{dataframe,}

\item{HospNum_Id}{Patient ID}

\item{Endo_ResultPerformed}{Date of the Endoscopy}
}
\description{
This determines the time difference between each test for a patient in days.
}
\examples{
em<-SurveillanceTimeByRow(Myendo,'HospitalNumber','Dateofprocedure')
}
\keyword{Surveillance}
