% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{SurveySankey}
\alias{SurveySankey}
\title{SurveySankey}
\usage{
SurveySankey(dfw, y)
}
\arguments{
\item{dfw}{the dataframe extracted using the standard cleanup scripts}

\item{y}{the column containing the test like ProcPerformed for example}
}
\description{
This creates a Sankey plot to see the order of tests for all patients:
dfw is the dataframe, y is the value of in this case
the procedure type (eg EMR,
 radiofrequency ablation for Barrett's but can be
 any description of a procedure you desire)
 Note the Hospital Number column MUST be called PatientID.
}
\keyword{Sankey}
