% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{SurveySankey}
\alias{SurveySankey}
\title{Create a Sankey plot for patient flow}
\usage{
SurveySankey(dfw, ProcPerformedColumn, PatientID)
}
\arguments{
\item{dfw}{the dataframe extracted using the standard cleanup scripts}

\item{ProcPerformedColumn}{the column containing the test like P
rocPerformed for example}

\item{PatientID}{the column containing the patients unique identifier
eg hostpital number}
}
\description{
The purpose of the function is to provide a Sankey plot 
which allows the analyst to see the proportion
of patients moving from one state (in this case type of Procedure) to
another. This allows us to see for example how many EMRs are done after
RFA.
}
\examples{
names(Myendo)[names(Myendo) == "HospitalNumber"] <- "PatientID"
gg <- SurveySankey(Myendo, "ProcedurePerformed", "PatientID")

}
\seealso{
Other Patient Flow functions: 
\code{\link{PatientFlowIndividual}()}
}
\concept{Patient Flow functions}
\keyword{Sankey}
