% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{TermStandardLocation}
\alias{TermStandardLocation}
\title{Standardises the location of biopsies
This assess the size of the biopies take}
\usage{
TermStandardLocation(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{y}{Column describing the Macroscopic sample from histology}
}
\description{
Standardises the location of biopsies
This assess the size of the biopies take
}
\keyword{Withdrawal}
