% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{TermStandardLocation}
\alias{TermStandardLocation}
\title{TermStandardLocation}
\usage{
TermStandardLocation(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{Column describing the Macroscopic sample from histology}
}
\description{
Standardises the location of biopsies by cleaning up the common typos and
abbreviations that are commonly used in free text of pathology reports
}
\examples{
Histoltree <-list("Hospital Number:","Patient Name:",
"General Practitioner:","Date received:","Clinical Details",
 "Nature of specimen","Histology","Diagnosis",""
)

for (i in 1:(length(Histoltree) - 1)) {
 Mypath <-Extractor(
     Mypath, "PathReportWhole",as.character(Histoltree[i]),
     as.character(Histoltree[i + 1]),
     gsub(" ", "", as.character(Histoltree[i]))
   )
}
names(Mypath)[names(Mypath) == 'Datereceived'] <- 'Dateofprocedure'
Mypath$Dateofprocedure <- as.Date(Mypath$Dateofprocedure)
f<-TermStandardLocation(Mypath,'Histology')
}
\keyword{Withdrawal}
