% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{TermStandardLocation}
\alias{TermStandardLocation}
\title{Standardises the location of biopsies
This assess the location of the biopies taken from the histopathology specimens.}
\usage{
TermStandardLocation(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{Column describing the Macroscopic sample from histology}
}
\description{
Standardises the location of biopsies
This assess the location of the biopies taken from the histopathology specimens.
}
\examples{
v<-MyColonpath
Histoltree<-list('Clinical Details','Nature of specimen','Histology','Diagnosis','')
v<-ChopperNewLines(v,'Histo_ResultText')
for(i in 1:(length(Histoltree)-1)) {
 v<-Extractor(v,'Histo_ResultText',as.character(Histoltree[i]),
 as.character(Histoltree[i+1]),gsub(' ','',as.character(Histoltree[i])))
}
f<-TermStandardLocation(v,'Histo_ResultText')
}
\keyword{Withdrawal}
