% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EndoMineR.R
\name{TermStandardLocation}
\alias{TermStandardLocation}
\title{TermStandardLocation}
\usage{
TermStandardLocation(x, SampleLocation)
}
\arguments{
\item{x}{The dataframe}

\item{SampleLocation}{Column describing the Macroscopic sample from histology}
}
\description{
Standardises the location of biopsies by cleaning up the common typos and 
abbreviations that are commonly used in free text of pathology reports
}
\examples{
f<-TermStandardLocation(Mypath,'Histology')
}
\keyword{Withdrawal}
