% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{textPrep}
\alias{textPrep}
\title{textPrep function}
\usage{
textPrep(inputText, delim, NegEx = c("TRUE", "FALSE"),
  Extractor = c("1", "2"), ExtractPOS = c("1", "2"))
}
\arguments{
\item{inputText}{The relevant pathology text column}

\item{delim}{the delimitors so the extractor can be used}

\item{NegEx}{parameter to say whether the NegativeRemove function used.}

\item{Extractor}{this states which Extractor you want to use. 1 is 
Extractor 1 (for uniformly ordered headers), 2 is Extractor2 for
text when headers are sometimes missing}
}
\value{
This returns a string vector.
}
\description{
This function prepares the data by cleaning 
punctuation, checking spelling against the lexicons, mapping terms
accorsing to the lexicons, removing negative expressions
and lower casing everything. It contains several of the other functions
in the package for ease of use. The user can decide whether to also include
POS tagging and Negative removal as well as which extractor. By default the
extractor called 'Extractor' (which assumes all headers are present in the
same order in each text entry) is used. Also by default the negative phrases
are removed and POS tagging is not performed.
}
\examples{
mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
CleanResults<-textPrep(PathDataFrameFinal$PathReportWhole,mywords,NegEx="TRUE",Extractor="1",ExtractPOS="2")
}
\keyword{Find}
\keyword{and}
\keyword{replace}
