% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_CleanUp.R
\name{textPrep}
\alias{textPrep}
\title{Combine all the text cleaning and extraction functions into one}
\usage{
textPrep(inputText, delim)
}
\arguments{
\item{inputText}{The relevant pathology text column}

\item{delim}{the delimitors so the extractor can be used}
}
\value{
This returns a string vector.
}
\description{
This function prepares the data by cleaning 
punctuation, checking spelling against the lexicons, mapping terms
according to the lexicons and lower casing everything. 
It contains several of the other functions
in the package for ease of use.
}
\examples{
mywords<-c("Hospital Number","Patient Name:","DOB:","General Practitioner:",
"Date received:","Clinical Details:","Macroscopic description:",
"Histology:","Diagnosis:")
CleanResults<-textPrep(PathDataFrameFinal$PathReportWhole,mywords)
}
\seealso{
Other NLP - Text Cleaning and Extraction: 
\code{\link{ColumnCleanUp}()},
\code{\link{DictionaryInPlaceReplace}()},
\code{\link{Extractor}()},
\code{\link{NegativeRemove}()}
}
\concept{NLP - Text Cleaning and Extraction}
\keyword{cleaning}
\keyword{text}
