% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Set the publication theme for all the ggplots}
\usage{
theme_Publication(base_size = 14, base_family = "Helvetica")
}
\arguments{
\item{base_size}{the base size}

\item{base_family}{the base family}
}
\description{
This standardises the theme for any ggplot plot produced.
If you do use it, like all ggplots it can be extended using the 
"+" to add whatever else is necessary
}
\examples{
# None needed
}
\seealso{
Other Data Presentation helpers: \code{\link{EndoBasicGraph}},
  \code{\link{scale_colour_Publication}},
  \code{\link{scale_fill_Publication}}
}
\concept{Data Presentation helpers}
\keyword{ggplot}
\keyword{themes}
