% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{WordsToNumbers}
\alias{WordsToNumbers}
\title{Convetr words to numbers especially for the histopathology text}
\usage{
WordsToNumbers()
}
\description{
This function converts words to numbers.
}
\seealso{
Other NLP - Lexicons: \code{\link{BiopsyIndex}},
  \code{\link{EventList}}, \code{\link{GISymptomsList}},
  \code{\link{HistolType}},
  \code{\link{LocationListLower}},
  \code{\link{LocationListUniversal}},
  \code{\link{LocationListUpper}},
  \code{\link{LocationList}}, \code{\link{RFACath}}
}
\concept{NLP - Lexicons}
\keyword{Event}
\keyword{extraction}
