% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lexicons.R
\name{WordsToNumbers}
\alias{WordsToNumbers}
\title{Words to numbers}
\usage{
WordsToNumbers()
}
\description{
This function converts words to numbers.
}
\examples{
# unique(unlist(EventList(), use.names = FALSE))

}
\keyword{Event}
\keyword{extraction}
